Calculate the generalized membership matrix (spatial version) with a noise cluster

Calculate the generalized membership matrix (spatial version) with a noise cluster

calcSFGCMBelongMatrixNoisy(
  centers,
  data,
  wdata,
  m,
  alpha,
  beta,
  delta,
  sigmas,
  wsigmas
)

Arguments

centers

A matrix representing the centers of the clusters with p columns and k rows

data

A matrix representing the observed data with n rows and p columns

wdata

A matrix representing the lagged observed data with n rows and p columns

m

A float representing the fuzziness degree

alpha

A float representing the weight of the space in the analysis (0 is a typical fuzzy-c-mean algorithm, 1 is balanced between the two dimensions, 2 is twice the weight for space)

beta

A float for the beta parameter (control speed convergence and classification crispness)

delta

A float, the value set for delta by the user

sigmas

A numeric vector for calculating the robust version of the FCM. Filled with ones if the classical version is required

wsigmas

Same as sigmas, but calculated on the spatially lagged dataset

Value

A n * k matrix representing the belonging probabilities of each observation to each cluster