Diversity index

Calculate the diversity (or entropy) index.

calcUncertaintyIndex(belongmatrix)

Arguments

belongmatrix

A membership matrix

Value

A vector with the values of the diversity (entropy) index

Details

The diversity (or entropy) index (Theil 1972) is calculated for each observation an varies between 0 and 1. When the value is close to 0, the observation belong to only one cluster (as in hard clustering). When the value is close to 1, the observation is undecided and tends to belong to each cluster. Values above 0.9 should be investigated. The formula is:

$$H2_{i} = \frac{-\sum[u_{ij}\ln(u_{ij})]}{\ln(k)}$$

with i and observation, j a cluster, k the number of clusters and u the membership matrix.

It is a simplified formula because the sum of each row of a membership matrix is 1.

References

Theil H (1972). Statistical decomposition analysis; with applications in the social and administrative sciences. North-Holland.

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- sf::st_drop_geometry(LyonIris[AnalysisFields])
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%[1] "criterion reached"
calcUncertaintyIndex(result$Belongings)
#>   [1] 0.63999705 0.25995148 0.28100726 0.67372988 0.24698005 0.20287281
#>   [7] 0.22920829 0.49136146 0.37509224 0.54880172 0.23763794 0.73505408
#>  [13] 0.52836440 0.70449169 0.62939082 0.48849316 0.49818774 0.68584038
#>  [19] 0.63433485 0.90204583 0.43547377 0.60061878 0.77911903 0.42795603
#>  [25] 0.32326842 0.42658930 0.83548460 0.65855008 0.29299921 0.56530151
#>  [31] 0.19687400 0.79867578 0.91809690 0.71216413 0.68334067 0.82014951
#>  [37] 0.27079624 0.41118365 0.53644132 0.92162006 0.63871103 0.29848558
#>  [43] 0.37639628 0.58260096 0.74826932 0.19012582 0.46836234 0.61313075
#>  [49] 0.75441085 0.66296378 0.36297097 0.45493928 0.87314229 0.30687876
#>  [55] 0.86108152 0.99516439 0.78040724 0.39927415 0.20069281 0.69717175
#>  [61] 0.42632436 0.59576850 0.22949737 0.94452354 0.45995276 0.21064821
#>  [67] 0.25420445 0.27083827 0.45511480 0.35548075 0.20346723 0.52940085
#>  [73] 0.60703188 0.12981092 0.46718477 0.65440730 0.34408303 0.51994274
#>  [79] 0.74948590 0.74344292 0.64913880 0.65518589 0.34056658 0.38145086
#>  [85] 0.42835266 0.42794728 0.56522872 0.19511832 0.64344967 0.83963531
#>  [91] 0.55135403 0.68399455 0.78068437 0.96030198 0.68680393 0.52483693
#>  [97] 0.61528007 0.89243094 0.73362806 0.89460201 0.29481398 0.44583839
#> [103] 0.56893968 0.36402745 0.96992041 0.68681802 0.66985023 0.24620345
#> [109] 0.35514022 0.80828043 0.60715788 0.80496139 0.15722869 0.68373517
#> [115] 0.16049743 0.63668177 0.63830838 0.64591735 0.34189931 0.17128686
#> [121] 0.80555679 0.48388480 0.39097125 0.13815375 0.71904728 0.80004610
#> [127] 0.93187629 0.59919482 0.62912553 0.22042825 0.49741999 0.55513250
#> [133] 0.19902540 0.72933736 0.30477712 0.39321862 0.49549060 0.68812454
#> [139] 0.24116097 0.36062956 0.17694083 0.64249741 0.75425743 0.39443464
#> [145] 0.63656364 0.67861771 0.71572530 0.40503197 0.34696990 0.41365313
#> [151] 0.58849544 0.18219433 0.86246243 0.49035326 0.44632114 0.85527812
#> [157] 0.39355188 0.70828603 0.74748677 0.49442396 0.41367384 0.46776680
#> [163] 0.24705395 0.55433715 0.21638273 0.50018825 0.88429676 0.55275804
#> [169] 0.60622269 0.25870857 0.33721091 0.69621844 0.64238839 0.50882487
#> [175] 0.80724859 0.67551320 0.65474995 0.80709287 0.81754557 0.30540077
#> [181] 0.24482740 0.65675192 0.64839260 0.75038135 0.73657138 0.57605694
#> [187] 0.61378030 0.74450023 0.65326913 0.95247575 0.39278937 0.89136673
#> [193] 0.38650927 0.74744736 0.52535758 0.32360317 0.14742754 0.82598467
#> [199] 0.22250491 0.43432598 0.36144416 0.82525395 0.90159834 0.22408700
#> [205] 0.60190522 0.60944545 0.82583254 0.67584668 0.77167818 0.18460319
#> [211] 0.76548885 0.40493026 0.71789899 0.50446724 0.87375363 0.55626388
#> [217] 0.67381066 0.52297336 0.25606242 0.17446410 0.11294278 0.63613981
#> [223] 0.14199450 0.34710313 0.57849122 0.24127784 0.19494958 0.75448550
#> [229] 0.62127962 0.21064136 0.52084589 0.37722274 0.72215406 0.61705272
#> [235] 0.60762857 0.35833990 0.51021346 0.36551073 0.65301499 0.68837279
#> [241] 0.40773873 0.35094126 0.16325312 0.20092113 0.93918967 0.58000896
#> [247] 0.55342430 0.61359045 0.11011708 0.87681965 0.60908869 0.59194533
#> [253] 0.67363257 0.65098265 0.82043620 0.56415020 0.82657433 0.84768223
#> [259] 0.45166367 0.54982272 0.59715599 0.07298928 0.44628918 0.58794694
#> [265] 0.53284656 0.71197558 0.86047379 0.19413038 0.27744979 0.30487034
#> [271] 0.52391631 0.34762084 0.24810142 0.84067708 0.30844689 0.30771623
#> [277] 0.52932935 0.06712005 0.22476936 0.72016803 0.69232958 0.47292254
#> [283] 0.78612489 0.48738976 0.31085872 0.15623588 0.25824529 0.36645703
#> [289] 0.78880811 0.86067676 0.75038691 0.77005208 0.56243148 0.70554397
#> [295] 0.76528311 0.19863093 0.89083153 0.47294094 0.35329394 0.31079501
#> [301] 0.75136879 0.57678106 0.91386334 0.74593230 0.55573565 0.24307961
#> [307] 0.58810754 0.34522675 0.23839405 0.11864142 0.53850990 0.75228766
#> [313] 0.49265593 0.61382038 0.23602619 0.21253738 0.16588371 0.62073980
#> [319] 0.85086667 0.38151352 0.40293916 0.04526931 0.64400771 0.59418154
#> [325] 0.56390752 0.65621313 0.62144312 0.17113875 0.41443003 0.72585319
#> [331] 0.54522491 0.33539359 0.44062871 0.02765532 0.54145142 0.83470544
#> [337] 0.81058828 0.74330564 0.57762167 0.84763970 0.25230715 0.71713147
#> [343] 0.27494057 0.23129590 0.44835385 0.19747374 0.93577215 0.54911774
#> [349] 0.41495258 0.85306499 0.42368349 0.09499536 0.69447987 0.34096107
#> [355] 0.17779529 0.59325705 0.55429558 0.62375406 0.77122541 0.72707187
#> [361] 0.48015747 0.22302447 0.42632756 0.48192233 0.87958486 0.75399880
#> [367] 0.48401529 0.45867730 0.67989832 0.07832338 0.76710720 0.87722063
#> [373] 0.38715485 0.20924914 0.57480981 0.85265338 0.93285734 0.63089812
#> [379] 0.37505827 0.56677840 0.63512323 0.57561993 0.25199560 0.21087705
#> [385] 0.59499479 0.43768974 0.71441064 0.39271499 0.70982748 0.69690539
#> [391] 0.61639868 0.62830925 0.33520110 0.62449423 0.53998966 0.80520707
#> [397] 0.87417959 0.81034573 0.49333841 0.65020565 0.66980833 0.49704080
#> [403] 0.52618106 0.60548776 0.63574829 0.74578587 0.58844824 0.86041896
#> [409] 0.08079351 0.65100398 0.29324333 0.90550726 0.63368476 0.54237055
#> [415] 0.50949985 0.40638765 0.77675908 0.49008508 0.65161604 0.69096258
#> [421] 0.19725232 0.90551761 0.67176112 0.83398189 0.20347781 0.56316491
#> [427] 0.21298574 0.83754625 0.62953207 0.49568432 0.55600610 0.82276713
#> [433] 0.54651904 0.66785115 0.85700891 0.08130837 0.48476516 0.49455597
#> [439] 0.34497453 0.74626063 0.24858614 0.67085884 0.56384234 0.23309366
#> [445] 0.30117724 0.59011359 0.90037076 0.67398958 0.41064137 0.58362468
#> [451] 0.66569804 0.56671081 0.35279673 0.58293969 0.42344330 0.68336984
#> [457] 0.29991018 0.31154775 0.47564679 0.81125448 0.47480646 0.59394393
#> [463] 0.18726292 0.68789835 0.55565439 0.25505315 0.26670774 0.81713635
#> [469] 0.92304155 0.88432611 0.90103052 0.86788325 0.84865373 0.88804387
#> [475] 0.81356971 0.82302881 0.91157780 0.77278290 0.86999233 0.36622636
#> [481] 0.62968961 0.64444227 0.52121518 0.63094981 0.90742519 0.81043819
#> [487] 0.56396035 0.79643762 0.84769152 0.76417561 0.25233566 0.84910348
#> [493] 0.83846762 0.70986052 0.52920761 0.78241297 0.62867153 0.84502006
#> [499] 0.13210770 0.89569990 0.47435241 0.76233418 0.52843800 0.71545327
#> [505] 0.48014504 0.70699499