Local Moran I for raster

Calculate the Local Moran I for a numeric raster

calc_local_moran_raster(rast, window)

Arguments

rast
A RasterLayer or a matrix

window
The window defining the neighbour weights

Value

A RasterLayer or a matrix depending on the input with the local Moran I values

Examples

data("Arcachon")
rast <- Arcachon[[1]]
w <- matrix(1, nrow = 3, ncol = 3)
calc_local_moran_raster(rast, w)

#> class      : RasterLayer 
#> dimensions : 347, 471, 163437  (nrow, ncol, ncell)
#> resolution : 30.00764, 30.04266  (x, y)
#> extent     : 360281.8, 374415.4, 6396254, 6406678  (xmin, xmax, ymin, ymax)
#> crs        : +proj=lcc +lat_0=46.5 +lon_0=3 +lat_1=49 +lat_2=44 +x_0=700000 +y_0=6600000 +ellps=GRS80 +units=m +no_defs 
#> source     : memory
#> names      : blue 
#> values     : -4.92144, 43.93583  (min, max)
#>