Spatial consistency index

Calculate a spatial consistency index

spConsistency(object, nblistw = NULL, window = NULL, nrep = 999, adj = FALSE)

Arguments

object
A FCMres object, typically obtained from functions CMeans, GCMeans, SFCMeans, SGFCMeans. Can also be a simple membership matrix.

nblistw
A list.w object describing the neighbours typically produced by the spdep package. Required if data is a dataframe, see the parameter window if you use a list of rasters as input. Can also be NULL if object is a FCMres object.

window
if rasters were used for the classification, the window must be specified instead of a list.w object. Can also be NULL if object is a FCMres object.

nrep
An integer indicating the number of permutation to do to simulate spatial randomness. Note that if rasters are used, each permutation can be very long.

adj
A boolean indicating if the adjusted version of the indicator must be calculated when working with rasters. When working with vectors, see the function adjustSpatialWeights to modify the list.w object.

Value

A named list with

Mean : the mean of the spatial consistency index

prt05 : the 5th percentile of the spatial consistency index

prt95 : the 95th percentile of the spatial consistency index

samples : all the value of the spatial consistency index

sum_diff : the total sum of squarred difference between observations and their neighbours

Details

This index is experimental, it aims to measure how much a clustering solution is spatially consistent. A classification is spatially inconsistent if neighbouring observation do not belong to the same group. See detail for a description of its calculation

The total spatial inconsistency (*Scr*) is calculated as follow

$$isp = \sum_{i}\sum_{j}\sum_{k} (u_{ik} - u_{jk})^{2} * W_{ij}$$

With U the membership matrix, i an observation, k the neighbours of i and W the spatial weight matrix This represents the total spatial inconsistency of the solution (true inconsistency) We propose to compare this total with simulated values obtained by permutations (simulated inconsistency). The values obtained by permutation are an approximation of the spatial inconsistency obtained in a random context Ratios between the true inconsistency and simulated inconsistencies are calculated A value of 0 depict a situation where all observations are identical to their neighbours A value of 1 depict a situation where all observations are as much different as their neighbours that what randomness can produce A classification solution able to reduce this index has a better spatial consistency

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- LyonIris@data[AnalysisFields]
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)

#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%[1] "criterion reached"

spConsistency(result$Belongings, nblistw = Wqueen, nrep=50)

#> $Mean
#> [1] 0.1613223
#> 
#> $Median
#>       50% 
#> 0.1615995 
#> 
#> $prt05
#>        5% 
#> 0.1579733 
#> 
#> $prt95
#>       95% 
#> 0.1649317 
#> 
#> $samples
#>  [1] 0.1619087 0.1595134 0.1586840 0.1622750 0.1626248 0.1643219 0.1628961
#>  [8] 0.1625251 0.1624275 0.1605331 0.1595634 0.1637501 0.1603729 0.1614530
#> [15] 0.1572192 0.1665429 0.1590308 0.1619753 0.1577213 0.1599467 0.1597228
#> [22] 0.1630630 0.1607127 0.1618150 0.1647976 0.1613570 0.1614384 0.1650414
#> [29] 0.1623760 0.1609807 0.1600640 0.1664063 0.1621578 0.1584902 0.1617460
#> [36] 0.1576184 0.1582813 0.1588678 0.1619252 0.1606895 0.1620584 0.1625833
#> [43] 0.1612205 0.1620516 0.1605062 0.1630639 0.1583774 0.1621159 0.1588249
#> [50] 0.1624744
#> 
#> $sum_diff
#> [1] 56.95028
#>