Descriptive statistics by group

Calculate some descriptive statistics of each group

summarizeClusters(data, belongmatrix, weighted = TRUE, dec = 3, silent = TRUE)

Arguments

data

The original dataframe used for the classification

belongmatrix

A membership matrix

weighted

A boolean indicating if the summary statistics must use the membership matrix columns as weights (TRUE) or simply assign each observation to its most likely cluster and compute the statistics on each subset (FALSE)

dec

An integer indicating the number of digits to keep when rounding (default is 3)

silent

A boolean indicating if the results must be printed or silently returned

Value

A list of length k (the number of group). Each element of the list is a dataframe with summary statistics for the variables of data for each group

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- sf::st_drop_geometry(LyonIris[AnalysisFields])
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%[1] "criterion reached"
summarizeClusters(dataset, result$Belongings)
#> $Cluster_1
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   50.774 22.188 15.032 6.577      10.418  7.35   6.26    7.543     
#> Q10  52.327 25.373 15.677 8.3        12.27   9.04   7.615   8.575     
#> Q25  54.491 29.306 17.083 10.683     14.873  11.742 11.17   11.5      
#> Q50  57.363 32.274 18.397 15.09      17.69   15.68  15.03   14.42     
#> Q75  61.416 36.22  18.83  20.69      20.332  19     18.99   17.528    
#> Q90  63.95  39.38  19.3   30.08      23.25   22.07  24.3    22.19     
#> Q95  66.009 40.927 19.706 34.81      25.75   25.048 31.855  28.435    
#> Mean 57.9   32.354 17.862 16.893     17.715  15.523 15.828  15.376    
#> Std  4.825  6.163  1.488  8.583      5.119   5.616  7.809   7.087     
#>      Pct_brevet NivVieMed
#> Q5   9.512      14995.5  
#> Q10  11.84      15636    
#> Q25  16.72      18439    
#> Q50  22.04      20083    
#> Q75  29.458     22594.75 
#> Q90  37.18      25281    
#> Q95  42.988     27156.75 
#> Mean 23.659     20504.18 
#> Std  10.972     3923.102 
#> 
#> $Cluster_2
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   50.785 19.37  13.901 6.415      14.972  7.52   8.33    9.512     
#> Q10  52.017 20.909 14.379 7.605      17.93   8.805  11.63   12.185    
#> Q25  54.424 23.017 15.656 10.84      21.165  11.11  19.77   16.84     
#> Q50  56.822 25.493 16.382 13.865     24.34   13.865 26.56   22.44     
#> Q75  58.526 29.806 17.113 18.902     27.58   17.358 32.34   30.75     
#> Q90  62.581 34.201 18.543 25.565     30.69   20.845 37.93   34.03     
#> Q95  64.279 38.565 18.94  29.637     32.3    23.798 40.722  37.98     
#> Mean 56.838 26.764 16.405 15.657     24.231  14.43  25.816  23.423    
#> Std  4.158  6.059  1.451  7.57       6.243   5.197  9.809   9.678     
#>      Pct_brevet NivVieMed
#> Q5   18.433     12351.5  
#> Q10  23         12747    
#> Q25  31.57      13837    
#> Q50  38.1       15846    
#> Q75  45.01      18411    
#> Q90  49.545     21142    
#> Q95  52.93      23522    
#> Mean 37.943     16474.23 
#> Std  12.06      3567.886 
#> 
#> $Cluster_3
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   48.551 15.711 13.514 6.235      12.55   8.985  4.858   6.69      
#> Q10  50.324 18.539 13.846 7.36       14.63   11.11  5.96    7.73      
#> Q25  52.48  20.5   14.642 11.235     17.22   13.95  8.285   9.64      
#> Q50  54.719 23.642 15.413 15.51      19.57   17.61  11.51   12.435    
#> Q75  57.763 27.625 16.383 23.148     21.83   21.15  17.1    16.032    
#> Q90  61.059 32.539 18.056 31.885     24.815  24.35  23.265  21.48     
#> Q95  63.269 37.376 18.8   37.71      26.683  27.48  29.727  27.63     
#> Mean 55.108 24.737 15.632 17.998     19.506  17.699 13.653  13.759    
#> Std  4.541  6.195  1.559  9.568      4.746   5.786  7.999   7.443     
#>      Pct_brevet NivVieMed
#> Q5   11.755     15277    
#> Q10  14.775     16312    
#> Q25  19.875     19001.75 
#> Q50  24.94      21303    
#> Q75  32.12      23791    
#> Q90  38.875     25957    
#> Q95  45.01      28142    
#> Mean 26.582     21448.47 
#> Std  11.072     3975.251 
#> 
#> $Cluster_4
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   49.819 26.339 15.994 5.362      8.13    6.47   5.73    7.45      
#> Q10  51.262 28.538 17.05  6.835      10.06   8.35   6.34    8.35      
#> Q25  53.458 30.922 18.445 9.86       12.14   10.74  8.07    10.87     
#> Q50  55.402 35.39  19.213 13.39      14.57   13.37  11.18   12.82     
#> Q75  58.564 39.393 19.768 17.5       17.088  17.275 13.54   15.15     
#> Q90  60.895 43.041 20.106 25.47      19.475  20.7   17.62   18.31     
#> Q95  63.035 44.426 20.252 32.062     21.375  23.508 20.01   20.59     
#> Mean 56.017 35.278 18.82  14.972     14.548  14.04  11.76   13.345    
#> Std  4.145  6.265  1.41   7.915      4.589   5.32   5.898   5.667     
#>      Pct_brevet NivVieMed
#> Q5   7.15       17713    
#> Q10  8.27       19059    
#> Q25  9.89       21207    
#> Q50  13.235     23224.5  
#> Q75  16.57      25924    
#> Q90  23.28      28806.5  
#> Q95  30.435     31467.25 
#> Mean 14.852     23611.37 
#> Std  8.586      4071.16  
#> 
#> $Cluster_5
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   44.572 14.777 12.388 10.477     12.78   11.55  3.56    6.732     
#> Q10  45.801 15.696 12.898 13.68      14.425  13.05  4.26    7.04      
#> Q25  49.699 18.707 13.696 20.7       16.93   15.985 5.84    7.95      
#> Q50  52.268 22.007 14.551 28.51      18.82   18.995 7.51    9.72      
#> Q75  54.852 25.783 15.656 36.38      20.85   22.205 10.33   12.08     
#> Q90  58.225 29.328 16.598 41.84      22.445  26.875 14.42   15.475    
#> Q95  60.558 33.019 17.473 43.65      23.112  28.86  19.29   18.185    
#> Mean 52.281 22.539 14.761 27.924     18.604  19.353 8.875   10.912    
#> Std  4.8    5.74   1.567  10.409     3.805   5.702  5.533   6.035     
#>      Pct_brevet NivVieMed
#> Q5   9.68       18921.5  
#> Q10  11.405     20257    
#> Q25  14.71      22402    
#> Q50  18.37      24532.5  
#> Q75  23.34      28327    
#> Q90  29.79      30951    
#> Q95  33.245     33673    
#> Mean 19.877     25276.26 
#> Std  8.882      4514.339 
#>