Undecided observations

Identify the observation for which the classification is uncertain

undecidedUnits(belongmatrix, tol = 0.1, out = "character")

Arguments

belongmatrix
The membership matrix obtained at the end of the algorithm

tol
A float indicating the minimum required level of membership to be not considered as undecided

out
The format of the output vector. Default is "character". If "numeric", then the undecided units are set to -1.

Value

A vector indicating the most likely group for each observation or "Undecided" if the maximum probability for the observation does not reach the value of the tol parameter

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- LyonIris@data[AnalysisFields]
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)

#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%[1] "criterion reached"

undecidedUnits(result$Belongings, tol = 0.45)

#>   [1] "V3"        "V2"        "V3"        "V5"        "V4"        "V3"       
#>   [7] "V4"        "V4"        "V4"        "V3"        "V4"        "V2"       
#>  [13] "V2"        "V4"        "V2"        "V4"        "V3"        "Undecided"
#>  [19] "V2"        "Undecided" "V5"        "V2"        "V3"        "V5"       
#>  [25] "V4"        "V4"        "Undecided" "V4"        "V4"        "V2"       
#>  [31] "V3"        "V5"        "Undecided" "V2"        "V2"        "V3"       
#>  [37] "V4"        "V3"        "V2"        "Undecided" "V4"        "V4"       
#>  [43] "V4"        "V3"        "V3"        "V4"        "V5"        "V2"       
#>  [49] "V2"        "V2"        "V4"        "V4"        "Undecided" "V2"       
#>  [55] "V3"        "Undecided" "V2"        "V2"        "V4"        "V2"       
#>  [61] "V4"        "V2"        "V2"        "Undecided" "V2"        "V3"       
#>  [67] "V4"        "V4"        "V2"        "V3"        "V4"        "V4"       
#>  [73] "V3"        "V3"        "V2"        "V5"        "V3"        "V2"       
#>  [79] "V2"        "V2"        "V4"        "V2"        "V4"        "V4"       
#>  [85] "V4"        "V2"        "V4"        "V3"        "V3"        "Undecided"
#>  [91] "V2"        "V2"        "V2"        "Undecided" "V2"        "V4"       
#>  [97] "V2"        "Undecided" "V2"        "Undecided" "V3"        "V3"       
#> [103] "V4"        "V3"        "Undecided" "V4"        "V3"        "V4"       
#> [109] "V3"        "V5"        "V3"        "Undecided" "V2"        "V2"       
#> [115] "V2"        "V2"        "V2"        "V2"        "V3"        "V2"       
#> [121] "V4"        "V2"        "V3"        "V4"        "Undecided" "V2"       
#> [127] "Undecided" "V4"        "V2"        "V4"        "V4"        "V5"       
#> [133] "V4"        "V3"        "V4"        "V3"        "V4"        "V4"       
#> [139] "V2"        "V4"        "V4"        "V5"        "V2"        "V2"       
#> [145] "V2"        "V3"        "V4"        "V4"        "V3"        "V3"       
#> [151] "V4"        "V3"        "V4"        "V5"        "V2"        "Undecided"
#> [157] "V4"        "V2"        "V2"        "V4"        "V4"        "V4"       
#> [163] "V4"        "V4"        "V4"        "V4"        "Undecided" "V2"       
#> [169] "V4"        "V4"        "V4"        "V3"        "V3"        "V2"       
#> [175] "V3"        "Undecided" "V2"        "Undecided" "V2"        "V4"       
#> [181] "V4"        "V3"        "V3"        "V5"        "V2"        "V2"       
#> [187] "V2"        "V1"        "Undecided" "Undecided" "V1"        "Undecided"
#> [193] "V1"        "V1"        "V1"        "V5"        "V1"        "V4"       
#> [199] "V1"        "V1"        "V3"        "Undecided" "Undecided" "V1"       
#> [205] "V5"        "V1"        "Undecided" "Undecided" "Undecided" "V1"       
#> [211] "Undecided" "V1"        "V5"        "V5"        "Undecided" "V5"       
#> [217] "V5"        "V4"        "V1"        "V4"        "V1"        "V1"       
#> [223] "V3"        "V1"        "V2"        "V3"        "V1"        "V5"       
#> [229] "V5"        "V1"        "V5"        "V1"        "Undecided" "V2"       
#> [235] "V3"        "V1"        "V5"        "V3"        "V5"        "V1"       
#> [241] "V5"        "V4"        "V4"        "V1"        "Undecided" "V1"       
#> [247] "V1"        "V5"        "V1"        "Undecided" "V5"        "V1"       
#> [253] "V1"        "V5"        "V1"        "V5"        "Undecided" "Undecided"
#> [259] "V1"        "V5"        "V2"        "V4"        "V4"        "V5"       
#> [265] "V1"        "V5"        "V4"        "V3"        "V4"        "V3"       
#> [271] "V5"        "V1"        "V3"        "Undecided" "V5"        "V1"       
#> [277] "V1"        "V1"        "V5"        "Undecided" "V1"        "V1"       
#> [283] "V5"        "V1"        "V3"        "V1"        "V2"        "V1"       
#> [289] "V1"        "Undecided" "V5"        "V5"        "V3"        "V4"       
#> [295] "V5"        "V1"        "Undecided" "V5"        "V3"        "V1"       
#> [301] "V5"        "V5"        "Undecided" "V1"        "V1"        "V5"       
#> [307] "V5"        "V5"        "V1"        "V1"        "V1"        "V5"       
#> [313] "V1"        "V4"        "V1"        "V1"        "V1"        "V1"       
#> [319] "Undecided" "V4"        "V4"        "V1"        "V4"        "V5"       
#> [325] "V1"        "V1"        "V5"        "V1"        "V1"        "V5"       
#> [331] "V3"        "V1"        "V5"        "V3"        "V1"        "Undecided"
#> [337] "V1"        "V1"        "V5"        "V3"        "V1"        "V5"       
#> [343] "V1"        "V1"        "V5"        "V1"        "Undecided" "V1"       
#> [349] "V1"        "Undecided" "V3"        "V1"        "V5"        "V5"       
#> [355] "V1"        "V1"        "V1"        "V5"        "V1"        "V1"       
#> [361] "V3"        "V3"        "V1"        "V5"        "V2"        "V1"       
#> [367] "V1"        "V5"        "V5"        "V3"        "V3"        "Undecided"
#> [373] "V1"        "V1"        "V3"        "V1"        "Undecided" "V5"       
#> [379] "V1"        "V5"        "V5"        "V4"        "V1"        "V1"       
#> [385] "V1"        "V1"        "V5"        "V1"        "V5"        "V4"       
#> [391] "V5"        "V5"        "V1"        "V5"        "V5"        "V3"       
#> [397] "Undecided" "V2"        "V1"        "V4"        "V5"        "V5"       
#> [403] "V1"        "V1"        "V1"        "V4"        "V1"        "V4"       
#> [409] "V1"        "V1"        "V2"        "Undecided" "V5"        "V4"       
#> [415] "V5"        "V5"        "V1"        "V1"        "V4"        "V5"       
#> [421] "V3"        "Undecided" "V5"        "V1"        "V1"        "V4"       
#> [427] "V1"        "Undecided" "V1"        "V1"        "V5"        "Undecided"
#> [433] "V5"        "V5"        "Undecided" "V1"        "V5"        "V5"       
#> [439] "V4"        "Undecided" "V3"        "V4"        "V1"        "V4"       
#> [445] "V1"        "V2"        "Undecided" "V3"        "V3"        "V1"       
#> [451] "V5"        "V5"        "V1"        "V3"        "V5"        "V5"       
#> [457] "V1"        "V1"        "V1"        "V4"        "V1"        "V1"       
#> [463] "V1"        "V5"        "V4"        "V1"        "V1"        "V1"       
#> [469] "Undecided" "Undecided" "Undecided" "Undecided" "Undecided" "Undecided"
#> [475] "V1"        "Undecided" "Undecided" "V4"        "Undecided" "V4"       
#> [481] "V4"        "V4"        "V4"        "V4"        "Undecided" "V4"       
#> [487] "V4"        "V3"        "Undecided" "V3"        "V4"        "Undecided"
#> [493] "Undecided" "V5"        "V5"        "Undecided" "V3"        "V5"       
#> [499] "V4"        "Undecided" "V4"        "V4"        "V4"        "Undecided"
#> [505] "V4"        "V2"