Diversity index

Calculate the diversity (or entropy) index.

calcUncertaintyIndex(belongmatrix)

Arguments

belongmatrix

A membership matrix

Value

A vector with the values of the diversity (entropy) index

Details

The diversity (or entropy) index (Theil 1972) is calculated for each observation an varies between 0 and 1. When the value is close to 0, the observation belong to only one cluster (as in hard clustering). When the value is close to 1, the observation is undecided and tends to belong to each cluster. Values above 0.9 should be investigated. The formula is:

$$H2_{i} = \frac{-\sum[u_{ij}\ln(u_{ij})]}{\ln(k)}$$

with i and observation, j a cluster, k the number of clusters and u the membership matrix.

It is a simplified formula because the sum of each row of a membership matrix is 1.

References

Theil H (1972). Statistical decomposition analysis; with applications in the social and administrative sciences. North-Holland.

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- sf::st_drop_geometry(LyonIris[AnalysisFields])
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%[1] "criterion reached"
calcUncertaintyIndex(result$Belongings)
#>   [1] 0.63862619 0.26265559 0.27967544 0.67395874 0.24832393 0.20284986
#>   [7] 0.22982202 0.49112062 0.37522719 0.54966678 0.23723053 0.73736176
#>  [13] 0.53381647 0.70476866 0.62687410 0.48793314 0.49569697 0.68525758
#>  [19] 0.63415620 0.90181545 0.43527241 0.60097000 0.77918459 0.42857349
#>  [25] 0.32491462 0.42603037 0.83529627 0.65846779 0.29234022 0.56890987
#>  [31] 0.19665410 0.79901600 0.91849806 0.71448995 0.68604342 0.81972663
#>  [37] 0.27204462 0.41181747 0.53777870 0.92163645 0.63873041 0.29906576
#>  [43] 0.37637808 0.58133768 0.74842307 0.19016336 0.46642693 0.61321921
#>  [49] 0.75739388 0.66667500 0.36267081 0.45585170 0.87301330 0.30873868
#>  [55] 0.86097335 0.99517911 0.78080289 0.40396757 0.20037091 0.70088991
#>  [61] 0.42693846 0.59986760 0.23334283 0.94465270 0.45939693 0.21060011
#>  [67] 0.25350155 0.27047866 0.45675287 0.35579469 0.20286148 0.52875077
#>  [73] 0.60595123 0.12979090 0.46865520 0.65423763 0.34418970 0.52372955
#>  [79] 0.75050063 0.74514209 0.64843050 0.65856084 0.34202242 0.38308005
#>  [85] 0.42810906 0.42819669 0.56543394 0.19473076 0.64267617 0.83936519
#>  [91] 0.55144121 0.68746449 0.78256165 0.96045618 0.68989424 0.52434624
#>  [97] 0.61566315 0.89213368 0.73493259 0.89423792 0.29491451 0.44431660
#> [103] 0.56845165 0.36223184 0.96995368 0.68631710 0.66861969 0.24585877
#> [109] 0.35553397 0.80783992 0.60563712 0.80471063 0.15940325 0.68475957
#> [115] 0.16112084 0.63991274 0.63737920 0.64825740 0.34177597 0.17388686
#> [121] 0.80527776 0.48298936 0.39134084 0.13790375 0.71883615 0.80150322
#> [127] 0.93184170 0.59847573 0.63153894 0.21982823 0.49661702 0.55454953
#> [133] 0.19883430 0.72855526 0.30423522 0.39319707 0.49480190 0.68796123
#> [139] 0.24543842 0.35995946 0.17624517 0.64372270 0.75464601 0.40049290
#> [145] 0.63909693 0.67721624 0.71536483 0.40527638 0.34686326 0.41372173
#> [151] 0.58793512 0.18208542 0.86231487 0.49003124 0.44719890 0.85517701
#> [157] 0.39419575 0.70866002 0.75049350 0.49402623 0.41305797 0.46715464
#> [163] 0.24680015 0.55417856 0.21764677 0.50097542 0.88454937 0.54980767
#> [169] 0.60615685 0.25955855 0.33724774 0.69522418 0.64090749 0.51075413
#> [175] 0.80691473 0.67502543 0.65645799 0.80756507 0.81887160 0.30818618
#> [181] 0.24639815 0.65534460 0.64683810 0.74946524 0.73826803 0.58120017
#> [187] 0.61371896 0.74482124 0.65098178 0.95254909 0.39240830 0.89177811
#> [193] 0.38558444 0.74768065 0.52506598 0.32290194 0.14737434 0.82678148
#> [199] 0.22205040 0.43442932 0.36163089 0.82492649 0.90180253 0.22268752
#> [205] 0.60037043 0.60898195 0.82523932 0.67464052 0.77103873 0.18433206
#> [211] 0.76486932 0.40434027 0.71613606 0.50483030 0.87400039 0.55539555
#> [217] 0.67405059 0.52471554 0.25588381 0.17630818 0.11263310 0.63579481
#> [223] 0.14190819 0.34675927 0.57535642 0.24155367 0.19467949 0.75338300
#> [229] 0.62086521 0.21065370 0.52193728 0.37675537 0.72089441 0.61605970
#> [235] 0.60857341 0.35838404 0.51233118 0.36345808 0.65358588 0.68841911
#> [241] 0.40596125 0.35450002 0.16370405 0.20047707 0.93916253 0.57950720
#> [247] 0.55308054 0.61433847 0.10972012 0.87663820 0.60983570 0.59149240
#> [253] 0.67393545 0.64911757 0.82075231 0.56473769 0.82617713 0.84833222
#> [259] 0.45126319 0.54760529 0.59392465 0.07415152 0.44827751 0.58967367
#> [265] 0.53244283 0.71275764 0.86111405 0.19388496 0.27954343 0.30490156
#> [271] 0.52417410 0.34702622 0.24785588 0.84086399 0.30745594 0.30737699
#> [277] 0.52896431 0.06696668 0.22452174 0.71907622 0.69235177 0.47255845
#> [283] 0.78603989 0.48761251 0.31103275 0.15581397 0.25799977 0.36652605
#> [289] 0.78889975 0.86018965 0.75099018 0.77048246 0.56361397 0.70730932
#> [295] 0.76580532 0.19828840 0.89107845 0.47080728 0.35421901 0.31045705
#> [301] 0.75024555 0.57564836 0.91375633 0.74588988 0.55513047 0.24252765
#> [307] 0.58919862 0.34493210 0.23823730 0.11826371 0.53873569 0.75222326
#> [313] 0.49203488 0.61629615 0.23581210 0.21242091 0.16545518 0.62031180
#> [319] 0.85049649 0.38349152 0.40603250 0.04514923 0.64238729 0.59492254
#> [325] 0.56400211 0.65609974 0.61976657 0.17100431 0.41348201 0.72501681
#> [331] 0.54601601 0.33464503 0.44248106 0.02773941 0.54087008 0.83391162
#> [337] 0.81072228 0.74335955 0.57898132 0.84742059 0.25188365 0.71809704
#> [343] 0.27464884 0.23064806 0.44681840 0.19650869 0.93554276 0.54879891
#> [349] 0.41501587 0.85090009 0.42254677 0.09481561 0.69449254 0.34021436
#> [355] 0.17759217 0.59308664 0.55424428 0.62378258 0.77125829 0.72711773
#> [361] 0.48035091 0.22362123 0.42642572 0.47862135 0.88035123 0.75422859
#> [367] 0.48412990 0.45914526 0.68101485 0.07845078 0.76705581 0.87737724
#> [373] 0.38721312 0.20902681 0.57520429 0.85293937 0.93261023 0.63152897
#> [379] 0.37523210 0.56608438 0.63517978 0.57671784 0.25169277 0.21024275
#> [385] 0.59482488 0.43788584 0.71405142 0.39284697 0.70941755 0.69684497
#> [391] 0.61532413 0.62846695 0.33527511 0.62560642 0.54081355 0.80501830
#> [397] 0.87447523 0.81197549 0.49337403 0.65086533 0.67058822 0.49523216
#> [403] 0.52641632 0.60513461 0.63542506 0.74551479 0.58837857 0.86088419
#> [409] 0.08058016 0.65085304 0.29305082 0.90575790 0.63493943 0.54361594
#> [415] 0.50818645 0.40781758 0.77707307 0.48941709 0.65221880 0.69164882
#> [421] 0.19769954 0.90535012 0.66979713 0.83411472 0.20342581 0.56578669
#> [427] 0.21267067 0.83730816 0.62950992 0.49578511 0.55644642 0.82202428
#> [433] 0.54532918 0.66883445 0.85595901 0.08109416 0.48225074 0.49554894
#> [439] 0.34738218 0.74389166 0.24899512 0.67279981 0.56374850 0.23623590
#> [445] 0.30075835 0.58690276 0.89986385 0.67411736 0.40946925 0.58312881
#> [451] 0.66750176 0.56513729 0.35211636 0.58274792 0.42248832 0.68129972
#> [457] 0.29962967 0.31139184 0.47482699 0.81155043 0.47447681 0.59378625
#> [463] 0.18706640 0.68824867 0.55835227 0.25403179 0.26636022 0.81733092
#> [469] 0.92283397 0.88459555 0.90119657 0.86751107 0.84613227 0.88835876
#> [475] 0.81337312 0.82256433 0.91141319 0.77283370 0.86919988 0.36821575
#> [481] 0.63085099 0.64147413 0.52362280 0.63268807 0.90742668 0.81141384
#> [487] 0.56600629 0.79633589 0.84774569 0.76395924 0.25463282 0.84871374
#> [493] 0.83838920 0.71021666 0.52772434 0.78116733 0.62834819 0.84518284
#> [499] 0.13263163 0.89593213 0.47876457 0.76120455 0.53071104 0.71510703
#> [505] 0.48426277 0.70722662