Diversity index

Calculate the diversity (or entropy) index.

calcUncertaintyIndex(belongmatrix)

Arguments

belongmatrix
A membership matrix

Value

A vector with the values of the diversity (entropy) index

Details

The diversity (or entropy) index (Theil 1972) is calculated for each observation an varies between 0 and 1. When the value is close to 0, the observation belong to only one cluster (as in hard clustering). When the value is close to 1, the observation is undecided and tends to belong to each cluster. Values above 0.9 should be investigated. The formula is:

$$H2_{i} = \frac{-\sum[u_{ij}\ln(u_{ij})]}{\ln(k)}$$

with i and observation, j a cluster, k the number of clusters and u the membership matrix.

It is a simplified formula because the sum of each row of a membership matrix is 1.

References

Theil H (1972). Statistical decomposition analysis; with applications in the social and administrative sciences. North-Holland.

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- LyonIris@data[AnalysisFields]
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)

#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%[1] "criterion reached"

calcUncertaintyIndex(result$Belongings)

#>   [1] 0.63852707 0.26279273 0.27963958 0.67394354 0.24837901 0.20283445
#>   [7] 0.22985224 0.49111698 0.37523639 0.54980152 0.23722290 0.73749039
#>  [13] 0.53408170 0.70476761 0.62673487 0.48791870 0.49553259 0.68521881
#>  [19] 0.63414058 0.90179783 0.43521996 0.60098733 0.77919934 0.42860244
#>  [25] 0.32497993 0.42601677 0.83528427 0.65846369 0.29232630 0.56910503
#>  [31] 0.19663065 0.79903713 0.91852040 0.71461840 0.68618972 0.81969597
#>  [37] 0.27209770 0.41186578 0.53782433 0.92163689 0.63872953 0.29909517
#>  [43] 0.37638437 0.58124615 0.74843105 0.19017125 0.46618208 0.61322985
#>  [49] 0.75754404 0.66687385 0.36266558 0.45588729 0.87300135 0.30884423
#>  [55] 0.86096867 0.99518006 0.78082500 0.40419300 0.20036609 0.70108270
#>  [61] 0.42696104 0.60009321 0.23354398 0.94466224 0.45934402 0.21062927
#>  [67] 0.25348582 0.27047378 0.45684990 0.35581362 0.20284844 0.52873090
#>  [73] 0.60587637 0.12979546 0.46870974 0.65421638 0.34417636 0.52393031
#>  [79] 0.75056023 0.74523571 0.64840754 0.65873481 0.34207780 0.38314449
#>  [85] 0.42810592 0.42819695 0.56543828 0.19474102 0.64263869 0.83935169
#>  [91] 0.55142080 0.68764371 0.78266286 0.96046522 0.69006356 0.52433169
#>  [97] 0.61568624 0.89211758 0.73498689 0.89419600 0.29496760 0.44430841
#> [103] 0.56843626 0.36210697 0.96995578 0.68627981 0.66855372 0.24585346
#> [109] 0.35554875 0.80776518 0.60555164 0.80470645 0.15950963 0.68481632
#> [115] 0.16114033 0.64005590 0.63730875 0.64835713 0.34176100 0.17401315
#> [121] 0.80526767 0.48295182 0.39135390 0.13790130 0.71882446 0.80158769
#> [127] 0.93184051 0.59845277 0.63164568 0.21981451 0.49659327 0.55443299
#> [133] 0.19883500 0.72850349 0.30422267 0.39322186 0.49478187 0.68793825
#> [139] 0.24565188 0.35994414 0.17622918 0.64382471 0.75466926 0.40081017
#> [145] 0.63923751 0.67714309 0.71533827 0.40529153 0.34684428 0.41376421
#> [151] 0.58791731 0.18206778 0.86230896 0.48996929 0.44725600 0.85519172
#> [157] 0.39422414 0.70866471 0.75063428 0.49401633 0.41304183 0.46713716
#> [163] 0.24679722 0.55416954 0.21769776 0.50100825 0.88456791 0.54965417
#> [169] 0.60615416 0.25959867 0.33726044 0.69520761 0.64080704 0.51086388
#> [175] 0.80692037 0.67499227 0.65653267 0.80760147 0.81894994 0.30829479
#> [181] 0.24647017 0.65527876 0.64672454 0.74937022 0.73836782 0.58147990
#> [187] 0.61371501 0.74484957 0.65084811 0.95255120 0.39238988 0.89178643
#> [193] 0.38543587 0.74767726 0.52501663 0.32288091 0.14737861 0.82680056
#> [199] 0.22201913 0.43445957 0.36163282 0.82489296 0.90180051 0.22246983
#> [205] 0.60019465 0.60888455 0.82517545 0.67452497 0.77100508 0.18431654
#> [211] 0.76482133 0.40426787 0.71595539 0.50484626 0.87401047 0.55529137
#> [217] 0.67408667 0.52477737 0.25586149 0.17638722 0.11260149 0.63573013
#> [223] 0.14195898 0.34672650 0.57519510 0.24157993 0.19465314 0.75325033
#> [229] 0.62078666 0.21065206 0.52200972 0.37663228 0.72081758 0.61598005
#> [235] 0.60869493 0.35840522 0.51262290 0.36338060 0.65367335 0.68843991
#> [241] 0.40573978 0.35464217 0.16372950 0.20045029 0.93916318 0.57941398
#> [247] 0.55296277 0.61436575 0.10967871 0.87663342 0.60995988 0.59140430
#> [253] 0.67395789 0.64891747 0.82077772 0.56479365 0.82614394 0.84840134
#> [259] 0.45121832 0.54733258 0.59375474 0.07419871 0.44835086 0.58992957
#> [265] 0.53239266 0.71287275 0.86112826 0.19389776 0.27963326 0.30488967
#> [271] 0.52428596 0.34695861 0.24788068 0.84085479 0.30728747 0.30733304
#> [277] 0.52892060 0.06694398 0.22453309 0.71901275 0.69236494 0.47250242
#> [283] 0.78603968 0.48763611 0.31102810 0.15574150 0.25798149 0.36652887
#> [289] 0.78888231 0.86014654 0.75106122 0.77054928 0.56373780 0.70735468
#> [295] 0.76583486 0.19826668 0.89110774 0.47055533 0.35436489 0.31035412
#> [301] 0.75011118 0.57552933 0.91376063 0.74587500 0.55499146 0.24248253
#> [307] 0.58940586 0.34487911 0.23819275 0.11823452 0.53876730 0.75226270
#> [313] 0.49187817 0.61637538 0.23580686 0.21240180 0.16539752 0.62021712
#> [319] 0.85048668 0.38356510 0.40615558 0.04514298 0.64228627 0.59501685
#> [325] 0.56401264 0.65606099 0.61955808 0.17098982 0.41331656 0.72491989
#> [331] 0.54613873 0.33457306 0.44272941 0.02774318 0.54075007 0.83383154
#> [337] 0.81071788 0.74333822 0.57919433 0.84741488 0.25180329 0.71824629
#> [343] 0.27463306 0.23058148 0.44666303 0.19635821 0.93552252 0.54873987
#> [349] 0.41503968 0.85076374 0.42248917 0.09480192 0.69447140 0.34008445
#> [355] 0.17754320 0.59299114 0.55415570 0.62379023 0.77127067 0.72712856
#> [361] 0.48035350 0.22371600 0.42644615 0.47827011 0.88039280 0.75424062
#> [367] 0.48415381 0.45922605 0.68117091 0.07845208 0.76707882 0.87737465
#> [373] 0.38722783 0.20901849 0.57528000 0.85295870 0.93259065 0.63162868
#> [379] 0.37524974 0.56603075 0.63515074 0.57675387 0.25167901 0.21017659
#> [385] 0.59473053 0.43789618 0.71401497 0.39287244 0.70942701 0.69681935
#> [391] 0.61520003 0.62847479 0.33529710 0.62579378 0.54093179 0.80502300
#> [397] 0.87449421 0.81206223 0.49340367 0.65087428 0.67072790 0.49499949
#> [403] 0.52644396 0.60503149 0.63534738 0.74549099 0.58834886 0.86089266
#> [409] 0.08056330 0.65077591 0.29299769 0.90575929 0.63508319 0.54365890
#> [415] 0.50803830 0.40801060 0.77707260 0.48930184 0.65221916 0.69177774
#> [421] 0.19772289 0.90533542 0.66956786 0.83412534 0.20344037 0.56587075
#> [427] 0.21263687 0.83725211 0.62947694 0.49581000 0.55649592 0.82194583
#> [433] 0.54516120 0.66890738 0.85588206 0.08107742 0.48193342 0.49564367
#> [439] 0.34747788 0.74376267 0.24900856 0.67285468 0.56374124 0.23636250
#> [445] 0.30064062 0.58674051 0.89984145 0.67416191 0.40946429 0.58301402
#> [451] 0.66767518 0.56496437 0.35204397 0.58277690 0.42240252 0.68108382
#> [457] 0.29960320 0.31133523 0.47469124 0.81154736 0.47442251 0.59370768
#> [463] 0.18705174 0.68835705 0.55844433 0.25389533 0.26633111 0.81734824
#> [469] 0.92281977 0.88462538 0.90119369 0.86747940 0.84598649 0.88838656
#> [475] 0.81332323 0.82253468 0.91141314 0.77283202 0.86916353 0.36829112
#> [481] 0.63088640 0.64130781 0.52370756 0.63273725 0.90743053 0.81143835
#> [487] 0.56607538 0.79635198 0.84773750 0.76397449 0.25472219 0.84866828
#> [493] 0.83836704 0.71026335 0.52755472 0.78109832 0.62836356 0.84518911
#> [499] 0.13265643 0.89592970 0.47892905 0.76111648 0.53079101 0.71510771
#> [505] 0.48441768 0.70723420