Diversity index

Calculate the diversity (or entropy) index.

calcUncertaintyIndex(belongmatrix)

Arguments

belongmatrix

A membership matrix

Value

A vector with the values of the diversity (entropy) index

Details

The diversity (or entropy) index (Theil 1972) is calculated for each observation an varies between 0 and 1. When the value is close to 0, the observation belong to only one cluster (as in hard clustering). When the value is close to 1, the observation is undecided and tends to belong to each cluster. Values above 0.9 should be investigated. The formula is:

$$H2_{i} = \frac{-\sum[u_{ij}\ln(u_{ij})]}{\ln(k)}$$

with i and observation, j a cluster, k the number of clusters and u the membership matrix.

It is a simplified formula because the sum of each row of a membership matrix is 1.

References

Theil H (1972). Statistical decomposition analysis; with applications in the social and administrative sciences. North-Holland.

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- sf::st_drop_geometry(LyonIris[AnalysisFields])
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%[1] "criterion reached"
calcUncertaintyIndex(result$Belongings)
#>   [1] 0.66541084 0.22661942 0.30655918 0.67316652 0.22722991 0.20585742
#>   [7] 0.21770683 0.49042106 0.37408972 0.52870578 0.24390398 0.69590922
#>  [13] 0.43770553 0.69814030 0.67900681 0.49225200 0.54408838 0.68391309
#>  [19] 0.64425482 0.90804052 0.44461553 0.60235175 0.77612816 0.42095470
#>  [25] 0.29623822 0.43189226 0.83917579 0.65891869 0.29927608 0.50340199
#>  [31] 0.20335853 0.79298536 0.91167273 0.66113929 0.63984227 0.82760260
#>  [37] 0.24843037 0.40051471 0.52912369 0.91860591 0.63706933 0.28526583
#>  [43] 0.37002997 0.60700349 0.74606445 0.18920144 0.51301032 0.61676066
#>  [49] 0.69394294 0.59300807 0.36802110 0.44038557 0.87533326 0.28690470
#>  [55] 0.86296523 0.99491257 0.77636043 0.32867188 0.20589367 0.62515155
#>  [61] 0.41935025 0.52606948 0.17262048 0.94152177 0.48727463 0.21208184
#>  [67] 0.26161032 0.27236760 0.43820921 0.35141468 0.20968438 0.53941008
#>  [73] 0.62776615 0.13152536 0.45391479 0.65687368 0.34488838 0.46260857
#>  [79] 0.72072908 0.70103193 0.65913979 0.57420638 0.31995841 0.35315513
#>  [85] 0.42920678 0.43929432 0.55956158 0.20289675 0.65743137 0.84401040
#>  [91] 0.56467796 0.62112096 0.74828329 0.95741738 0.62087696 0.53238667
#>  [97] 0.61631338 0.89897611 0.71787496 0.90242824 0.29271669 0.47228153
#> [103] 0.57263319 0.40024425 0.96944357 0.69429440 0.68037788 0.25080898
#> [109] 0.35005313 0.81995001 0.63409104 0.80651592 0.13329570 0.66952463
#> [115] 0.16425974 0.58762867 0.66301923 0.61532682 0.34640831 0.13763302
#> [121] 0.80942865 0.51083628 0.38626360 0.14176698 0.71237903 0.77531164
#> [127] 0.93261023 0.60974325 0.59597697 0.22897791 0.50895577 0.57294896
#> [133] 0.20144332 0.73785576 0.31254204 0.39407615 0.50583789 0.68825905
#> [139] 0.17999049 0.36679332 0.18539418 0.62038041 0.74939464 0.29548527
#> [145] 0.59327022 0.70044801 0.71764940 0.39902401 0.35140696 0.41233705
#> [151] 0.59387294 0.18660878 0.86467094 0.50148577 0.44144492 0.85122756
#> [157] 0.38302417 0.71169494 0.68774203 0.50014952 0.42241324 0.47740229
#> [163] 0.25170501 0.55716174 0.19917519 0.47965262 0.87802856 0.58888449
#> [169] 0.60423515 0.24280286 0.33056072 0.70842610 0.66854616 0.48584876
#> [175] 0.81103057 0.68614506 0.63261427 0.79180628 0.79304368 0.26243415
#> [181] 0.21633020 0.67363996 0.67598775 0.76826955 0.70875346 0.48234262
#> [187] 0.61796657 0.73843562 0.68257295 0.95035018 0.39907312 0.88347894
#> [193] 0.40779642 0.74409909 0.53255546 0.33866669 0.14870672 0.81311353
#> [199] 0.23094419 0.43179145 0.36024853 0.82666067 0.89726787 0.25779786
#> [205] 0.63581193 0.62039482 0.83737291 0.69713864 0.78176258 0.18981728
#> [211] 0.77618746 0.41716869 0.75239757 0.50111362 0.86997909 0.57738902
#> [217] 0.66883055 0.49184516 0.26070586 0.14561935 0.11971090 0.64437177
#> [223] 0.14279168 0.35421149 0.63503167 0.23760155 0.20091940 0.77829417
#> [229] 0.63375366 0.21150718 0.50457470 0.39167476 0.74259101 0.63850872
#> [235] 0.58545886 0.35729862 0.46431392 0.40409932 0.63777473 0.68680969
#> [241] 0.44950121 0.29161885 0.15526917 0.20890543 0.93848576 0.59096352
#> [247] 0.56497769 0.60431535 0.11865424 0.87961991 0.59085643 0.60177320
#> [253] 0.66840918 0.68713311 0.81429670 0.55084258 0.83372967 0.83122426
#> [259] 0.45997992 0.59864530 0.64200373 0.05699309 0.41711933 0.54784110
#> [265] 0.54077626 0.69448199 0.85023141 0.19940987 0.24222138 0.30688202
#> [271] 0.51619120 0.35991275 0.25297458 0.83531838 0.33616722 0.31568512
#> [277] 0.53607221 0.07121675 0.23144385 0.73794994 0.69147763 0.48143062
#> [283] 0.78901690 0.48380015 0.31027167 0.16729553 0.27636032 0.36640024
#> [289] 0.78792611 0.86249902 0.73840467 0.75778689 0.53788110 0.67322091
#> [295] 0.75761369 0.20512418 0.88285484 0.52057558 0.33384570 0.32273473
#> [301] 0.77569036 0.60245570 0.91144656 0.74679336 0.57049727 0.25735482
#> [307] 0.56106888 0.35635224 0.24438317 0.12604236 0.53423948 0.74943663
#> [313] 0.51015077 0.57656201 0.23997864 0.21618417 0.17598061 0.63065360
#> [319] 0.85698948 0.35342086 0.34757765 0.04769423 0.65616200 0.57760598
#> [325] 0.56238337 0.65864275 0.65873878 0.17461128 0.43692006 0.74471749
#> [331] 0.52744814 0.34953724 0.40198527 0.02694297 0.55373057 0.84512277
#> [337] 0.80855707 0.74197805 0.54733645 0.85102269 0.26366781 0.69382722
#> [343] 0.28035567 0.24451399 0.48271897 0.22064005 0.94029748 0.55531938
#> [349] 0.41335074 0.88323223 0.44586891 0.09882571 0.69681630 0.36301001
#> [355] 0.18456175 0.60085534 0.56057831 0.62281491 0.77026279 0.72618144
#> [361] 0.47860407 0.21120679 0.42448815 0.55045415 0.86607155 0.74992934
#> [367] 0.48178606 0.45038164 0.65319309 0.07759746 0.76634174 0.87435391
#> [373] 0.38636961 0.21356044 0.56600786 0.84717379 0.93758735 0.61776609
#> [379] 0.37273017 0.58254648 0.63786550 0.55359228 0.25747091 0.22387654
#> [385] 0.60184528 0.43536373 0.72234955 0.39020815 0.71670063 0.69113627
#> [391] 0.64106003 0.62287640 0.33371956 0.59857295 0.52264071 0.80749231
#> [397] 0.86649247 0.78070030 0.49159088 0.63307896 0.65048635 0.53868059
#> [403] 0.52206803 0.61591082 0.64436723 0.74819427 0.59124276 0.85330989
#> [409] 0.08526994 0.65703024 0.31112289 0.89935558 0.60578022 0.51419306
#> [415] 0.53963567 0.37794769 0.77088622 0.50560484 0.63198985 0.67444894
#> [421] 0.19094365 0.90831709 0.71234154 0.83142092 0.20413580 0.50744918
#> [427] 0.22005358 0.84133869 0.62836978 0.49343337 0.54737524 0.83379797
#> [433] 0.57572337 0.64643243 0.86607630 0.08578439 0.54115308 0.47794023
#> [439] 0.30788640 0.77448304 0.24340167 0.63548332 0.56516138 0.18431752
#> [445] 0.31512349 0.63821179 0.90440920 0.66754898 0.43147595 0.59402816
#> [451] 0.63026432 0.60145397 0.36618374 0.58570248 0.44555496 0.72232726
#> [457] 0.30592081 0.31826779 0.49363487 0.80526472 0.48296272 0.59963323
#> [463] 0.19154474 0.67812367 0.51437899 0.27729366 0.27373725 0.81340464
#> [469] 0.92650138 0.87381926 0.89600546 0.87069327 0.88046739 0.88135304
#> [475] 0.81794955 0.83139137 0.91321261 0.77208743 0.88135744 0.33720796
#> [481] 0.61294027 0.67248745 0.48236199 0.59641373 0.90660770 0.79489341
#> [487] 0.53145259 0.79714628 0.84607297 0.76668136 0.21926319 0.85569229
#> [493] 0.83596353 0.70313223 0.56303099 0.79887883 0.63383757 0.84229776
#> [499] 0.12455143 0.88937696 0.39675754 0.76895662 0.49333952 0.72016973
#> [505] 0.40842886 0.70810044