Diversity index

Calculate the diversity (or entropy) index.

calcUncertaintyIndex(belongmatrix)

Arguments

belongmatrix

A membership matrix

Value

A vector with the values of the diversity (entropy) index

Details

The diversity (or entropy) index (Theil 1972) is calculated for each observation an varies between 0 and 1. When the value is close to 0, the observation belong to only one cluster (as in hard clustering). When the value is close to 1, the observation is undecided and tends to belong to each cluster. Values above 0.9 should be investigated. The formula is:

$$H2_{i} = \frac{-\sum[u_{ij}\ln(u_{ij})]}{\ln(k)}$$

with i and observation, j a cluster, k the number of clusters and u the membership matrix.

It is a simplified formula because the sum of each row of a membership matrix is 1.

References

Theil H (1972). Statistical decomposition analysis; with applications in the social and administrative sciences. North-Holland.

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- sf::st_drop_geometry(LyonIris[AnalysisFields])
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%[1] "criterion reached"
calcUncertaintyIndex(result$Belongings)
#>   [1] 0.66657883 0.22723572 0.30413962 0.67609626 0.22810883 0.20704518
#>   [7] 0.21807432 0.49042945 0.37406172 0.52321843 0.24332884 0.69612267
#>  [13] 0.44047299 0.69916890 0.67854379 0.49198158 0.54528585 0.68417356
#>  [19] 0.64433996 0.90808237 0.44836383 0.60220668 0.77492672 0.42335661
#>  [25] 0.29753427 0.43147680 0.83923357 0.65907836 0.29865774 0.50438060
#>  [31] 0.20434864 0.79341824 0.91164891 0.66154224 0.64025454 0.82795623
#>  [37] 0.24938858 0.40018492 0.53011345 0.91857760 0.63737013 0.28574109
#>  [43] 0.37023314 0.60820843 0.74628742 0.18902622 0.52178447 0.61639434
#>  [49] 0.69572462 0.59427901 0.36760440 0.44117340 0.87541864 0.28663994
#>  [55] 0.86292584 0.99489621 0.77632099 0.33087592 0.20537147 0.62686667
#>  [61] 0.41982277 0.52661211 0.17367303 0.94138784 0.48771970 0.21037739
#>  [67] 0.26091525 0.27204264 0.43769981 0.35172582 0.20905723 0.53881892
#>  [73] 0.62861927 0.13143018 0.45529774 0.65753590 0.34646840 0.46323594
#>  [79] 0.72112986 0.70196327 0.65858816 0.57622332 0.32101162 0.35458046
#>  [85] 0.42906523 0.43955474 0.56020509 0.20108133 0.65718332 0.84360377
#>  [91] 0.56549334 0.62259468 0.74864167 0.95737493 0.62186462 0.53196390
#>  [97] 0.61597918 0.89881843 0.71877007 0.90370730 0.29008893 0.46752913
#> [103] 0.57250919 0.40172436 0.96941770 0.69490139 0.68042125 0.25030871
#> [109] 0.35093249 0.82346752 0.63404300 0.80600271 0.13381723 0.66964436
#> [115] 0.16470704 0.58986768 0.66342407 0.61682667 0.34724848 0.13863683
#> [121] 0.80928289 0.50977911 0.38716743 0.14132052 0.71258703 0.77526253
#> [127] 0.93267817 0.60914913 0.59741930 0.22824441 0.50820187 0.57925854
#> [133] 0.20105168 0.73814047 0.31190888 0.39278141 0.50514685 0.68919482
#> [139] 0.18142219 0.36623960 0.18462719 0.61965518 0.74932124 0.29722853
#> [145] 0.59368986 0.70013248 0.71826467 0.39921865 0.35256923 0.41024636
#> [151] 0.59365634 0.18763488 0.86461258 0.50550639 0.44084339 0.84985433
#> [157] 0.38349051 0.71217234 0.69019181 0.49971998 0.42178550 0.47676632
#> [163] 0.25124801 0.55736645 0.19996054 0.48083143 0.87778189 0.58925699
#> [169] 0.60455979 0.24340898 0.33064666 0.70631071 0.66944203 0.48555101
#> [175] 0.80947315 0.68641980 0.63374940 0.79166398 0.79294009 0.26455128
#> [181] 0.21741780 0.67294538 0.67733960 0.77077447 0.70861683 0.48339112
#> [187] 0.61799923 0.73768045 0.68348557 0.95043324 0.39780340 0.88429509
#> [193] 0.41394941 0.74512935 0.53446735 0.33765002 0.14791556 0.81430879
#> [199] 0.23031757 0.42968048 0.36116868 0.82762609 0.89804580 0.26767244
#> [205] 0.64148809 0.62496178 0.83869996 0.69806690 0.78073217 0.18916757
#> [211] 0.77603288 0.41921878 0.75622745 0.50328141 0.87080021 0.58167579
#> [217] 0.66776803 0.49384002 0.26128895 0.14681544 0.12045829 0.64755717
#> [223] 0.13951887 0.35473546 0.63419530 0.23730926 0.20143041 0.78230995
#> [229] 0.63811587 0.21175472 0.50582507 0.39900964 0.74306364 0.63877207
#> [235] 0.58069769 0.35559898 0.45340316 0.40170253 0.63280680 0.68527814
#> [241] 0.45766562 0.29444449 0.15543808 0.20796808 0.93858161 0.59472911
#> [247] 0.57215393 0.60690800 0.11956600 0.87929445 0.58512207 0.60531678
#> [253] 0.66800524 0.69108862 0.81374940 0.54882152 0.83395434 0.82885951
#> [259] 0.46111652 0.60754397 0.64222788 0.05763791 0.41881443 0.53678641
#> [265] 0.54208636 0.68987871 0.85124235 0.19794983 0.24373042 0.30808601
#> [271] 0.51011855 0.36158780 0.25078744 0.83639946 0.34484386 0.31709981
#> [277] 0.53692904 0.07230397 0.23049518 0.73673860 0.69047714 0.48389120
#> [283] 0.78899839 0.48288050 0.31156326 0.17098490 0.27622726 0.36625617
#> [289] 0.78941828 0.86344249 0.73645048 0.75386651 0.53445648 0.67616003
#> [295] 0.75858686 0.20455260 0.88232734 0.52871306 0.32852829 0.32929136
#> [301] 0.78006604 0.60605031 0.91092608 0.74751736 0.57699143 0.25919272
#> [307] 0.55103833 0.36002154 0.24710557 0.12597803 0.53267287 0.74589608
#> [313] 0.51881863 0.57933591 0.23888536 0.21702434 0.17821001 0.63497904
#> [319] 0.85612410 0.35497371 0.35038030 0.04749345 0.65764839 0.57382905
#> [325] 0.56190389 0.66023555 0.66580110 0.17505636 0.44386534 0.74791050
#> [331] 0.52261720 0.35014923 0.39404341 0.02713246 0.55794262 0.84741894
#> [337] 0.80932719 0.74309530 0.53858027 0.85066311 0.26798870 0.68693939
#> [343] 0.27960392 0.24572088 0.48703836 0.22760779 0.94078661 0.55714035
#> [349] 0.41154601 0.88399674 0.44560987 0.09887050 0.69900304 0.37008473
#> [355] 0.18750618 0.60708020 0.56702366 0.62249169 0.76944887 0.72555641
#> [361] 0.47947199 0.20791292 0.42323219 0.55955677 0.86622217 0.74991322
#> [367] 0.48016339 0.44798778 0.64613546 0.07821132 0.76451660 0.87493492
#> [373] 0.38513768 0.21256203 0.56283345 0.84694325 0.93794985 0.61410039
#> [379] 0.37208667 0.58363938 0.64118110 0.55521129 0.25653459 0.22521893
#> [385] 0.60682295 0.43545254 0.72372594 0.38872211 0.71453282 0.69276904
#> [391] 0.64550883 0.62140921 0.33220403 0.59029217 0.51846510 0.80625605
#> [397] 0.86612498 0.78118696 0.48909359 0.63495846 0.64407357 0.54730483
#> [403] 0.52087531 0.62181302 0.64868552 0.74885391 0.59265698 0.85414347
#> [409] 0.08536534 0.66205220 0.31332203 0.89997952 0.60067151 0.51614371
#> [415] 0.54456988 0.37256494 0.77202094 0.51085630 0.63435959 0.66839567
#> [421] 0.19146242 0.90909138 0.71848416 0.83120504 0.20248035 0.51140991
#> [427] 0.22087858 0.84429061 0.62864416 0.49173981 0.54649352 0.83608859
#> [433] 0.58292423 0.64406794 0.86724429 0.08586182 0.55168390 0.47728495
#> [439] 0.30972025 0.77491185 0.24443171 0.63838463 0.56497371 0.18647215
#> [445] 0.32247099 0.63792426 0.90426548 0.66408882 0.42778026 0.59823418
#> [451] 0.62672744 0.60680975 0.36753086 0.58340238 0.44794841 0.72600432
#> [457] 0.30636166 0.32198783 0.49896294 0.80634740 0.48516130 0.60389248
#> [463] 0.19158663 0.67215675 0.51709011 0.28211020 0.27386857 0.81302544
#> [469] 0.92658754 0.87259973 0.89661923 0.87134973 0.88067026 0.88098753
#> [475] 0.82062625 0.83120934 0.91278967 0.77236727 0.88127340 0.33877074
#> [481] 0.61427721 0.67381247 0.48478278 0.59923402 0.90643118 0.79629099
#> [487] 0.53361977 0.79571944 0.84664169 0.76475035 0.22086816 0.85714923
#> [493] 0.83700585 0.70224779 0.56849461 0.79911394 0.63180025 0.84323345
#> [499] 0.12474754 0.89018069 0.40114624 0.77085235 0.49557079 0.71882578
#> [505] 0.41248388 0.70828014