Diversity index

Calculate the diversity (or entropy) index.

calcUncertaintyIndex(belongmatrix)

Arguments

belongmatrix

A membership matrix

Value

A vector with the values of the diversity (entropy) index

Details

The diversity (or entropy) index (Theil 1972) is calculated for each observation an varies between 0 and 1. When the value is close to 0, the observation belong to only one cluster (as in hard clustering). When the value is close to 1, the observation is undecided and tends to belong to each cluster. Values above 0.9 should be investigated. The formula is:

$$H2_{i} = \frac{-\sum[u_{ij}\ln(u_{ij})]}{\ln(k)}$$

with i and observation, j a cluster, k the number of clusters and u the membership matrix.

It is a simplified formula because the sum of each row of a membership matrix is 1.

References

Theil H (1972). Statistical decomposition analysis; with applications in the social and administrative sciences. North-Holland.

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- sf::st_drop_geometry(LyonIris[AnalysisFields])
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%[1] "criterion reached"
calcUncertaintyIndex(result$Belongings)
#>   [1] 0.66868252 0.22494235 0.30742505 0.67472280 0.22643734 0.20717876
#>   [7] 0.21700930 0.49009076 0.37395696 0.52291242 0.24387253 0.69270664
#>  [13] 0.43225906 0.69822145 0.68327985 0.49201894 0.54922858 0.68279477
#>  [19] 0.64621914 0.90874509 0.44699231 0.60276638 0.77478263 0.42168841
#>  [25] 0.29514037 0.43176573 0.83965154 0.65904734 0.29902916 0.49903117
#>  [31] 0.20483506 0.79250430 0.91112914 0.65605184 0.63655256 0.82860964
#>  [37] 0.24734632 0.39944661 0.53035367 0.91806360 0.63712510 0.28440071
#>  [43] 0.36935983 0.61027901 0.74609347 0.18895217 0.52306743 0.61701302
#>  [49] 0.68955888 0.58775496 0.36803019 0.43978633 0.87544264 0.28602933
#>  [55] 0.86314945 0.99488607 0.77611580 0.32483503 0.20582654 0.61993018
#>  [61] 0.41921495 0.52043358 0.16935710 0.94115869 0.49123167 0.21111870
#>  [67] 0.26141209 0.27203970 0.43693815 0.35143318 0.20948445 0.53964362
#>  [73] 0.63044522 0.13176798 0.45473154 0.65727475 0.34627595 0.45849574
#>  [79] 0.71768280 0.69704692 0.65939115 0.56677355 0.31928802 0.35194446
#>  [85] 0.42900561 0.44146181 0.55946485 0.20250550 0.65860474 0.84399833
#>  [91] 0.56748890 0.61665205 0.74605860 0.95714880 0.61498082 0.53254202
#>  [97] 0.61661438 0.89958961 0.71782831 0.90384504 0.29087180 0.47157163
#> [103] 0.57259066 0.40494023 0.96938875 0.69524045 0.68031797 0.25073191
#> [109] 0.35038424 0.82292569 0.63656842 0.80619710 0.13271723 0.66910191
#> [115] 0.16602153 0.58564408 0.66649538 0.61470471 0.34759591 0.13649920
#> [121] 0.80959584 0.51288792 0.38663229 0.14166085 0.71103802 0.77321305
#> [127] 0.93272725 0.61000466 0.59515589 0.22895726 0.50917875 0.57828809
#> [133] 0.20121972 0.73836815 0.31254638 0.39340984 0.50605744 0.68873236
#> [139] 0.17668600 0.36658115 0.18529091 0.61731810 0.74902748 0.28864080
#> [145] 0.59043199 0.70196097 0.71822289 0.39849427 0.35282141 0.41084577
#> [151] 0.59387019 0.18791027 0.86479621 0.50461212 0.44147911 0.84927572
#> [157] 0.38245712 0.71296513 0.68406851 0.50023971 0.42251088 0.47760365
#> [163] 0.25166593 0.55746739 0.19852662 0.47839031 0.87726954 0.59086060
#> [169] 0.60411368 0.24192017 0.32977565 0.70790524 0.67153715 0.48416287
#> [175] 0.81028539 0.68732045 0.63211234 0.78997336 0.79079349 0.26085546
#> [181] 0.21475411 0.67408604 0.67940192 0.77102493 0.70638914 0.47465815
#> [187] 0.61847033 0.73791158 0.68480672 0.95014656 0.39827459 0.88331432
#> [193] 0.41235711 0.74457993 0.53412811 0.34099830 0.14870774 0.81301842
#> [199] 0.23092887 0.43070106 0.36097445 0.82727145 0.89730437 0.26560635
#> [205] 0.64285989 0.62330615 0.83897069 0.69918112 0.78166223 0.18987371
#> [211] 0.77673065 0.41860208 0.75772281 0.50287548 0.86978406 0.58277329
#> [217] 0.66818131 0.49072123 0.26173330 0.14437025 0.12049884 0.64639046
#> [223] 0.14077168 0.35467536 0.63935419 0.23728033 0.20158095 0.78307651
#> [229] 0.63739031 0.21213991 0.50324434 0.39708086 0.74448243 0.64109761
#> [235] 0.57994415 0.35654440 0.45440619 0.40621440 0.63370748 0.68578319
#> [241] 0.45918067 0.28911215 0.15469471 0.20857121 0.93818601 0.59334098
#> [247] 0.56985822 0.60414447 0.11968060 0.87983727 0.58652446 0.60396674
#> [253] 0.66825957 0.69263773 0.81374063 0.54838027 0.83431296 0.82766900
#> [259] 0.46074516 0.60879484 0.64496306 0.05639592 0.41620305 0.53818840
#> [265] 0.54151099 0.69023087 0.85014682 0.19912646 0.24054204 0.30814845
#> [271] 0.51237738 0.36117521 0.25217513 0.83524867 0.34401438 0.31708551
#> [277] 0.53653900 0.07227516 0.23239824 0.73836561 0.69076482 0.48307140
#> [283] 0.78954647 0.48365032 0.31132426 0.17031961 0.27905090 0.36701656
#> [289] 0.78864573 0.86290637 0.73637573 0.75423383 0.53318906 0.67262671
#> [295] 0.75691858 0.20508563 0.88109635 0.53090330 0.32851132 0.32773639
#> [301] 0.78055247 0.60797219 0.91041540 0.74702071 0.57479196 0.26094168
#> [307] 0.55383547 0.35987779 0.24692618 0.12621744 0.53354649 0.74672988
#> [313] 0.51617429 0.57582493 0.23964785 0.21731433 0.17788637 0.63332932
#> [319] 0.85704957 0.35258019 0.34515487 0.04770905 0.65721933 0.57423443
#> [325] 0.56216698 0.65935043 0.66674446 0.17525556 0.44235556 0.74829079
#> [331] 0.52251781 0.35031106 0.39469158 0.02708588 0.55647713 0.84689990
#> [337] 0.80879012 0.74213269 0.53960375 0.85108952 0.26685746 0.68772590
#> [343] 0.28016267 0.24568295 0.48883745 0.22581934 0.94099773 0.55640681
#> [349] 0.41246189 0.88556833 0.44790993 0.09911395 0.69849956 0.36945527
#> [355] 0.18711978 0.60498561 0.56507623 0.62265760 0.76969884 0.72575622
#> [361] 0.47925878 0.20797010 0.42393558 0.56170438 0.86514636 0.74962937
#> [367] 0.48115477 0.45011832 0.64665166 0.07807247 0.76490088 0.87454200
#> [373] 0.38617228 0.21357110 0.56310358 0.84681452 0.93823761 0.61446275
#> [379] 0.37291320 0.58525947 0.64025736 0.55278919 0.25706050 0.22505242
#> [385] 0.60518753 0.43591636 0.72338101 0.38957524 0.71712908 0.69144198
#> [391] 0.64689853 0.62129774 0.33309157 0.59197747 0.51979614 0.80674119
#> [397] 0.86526023 0.77841095 0.49011425 0.63272417 0.64542719 0.54811493
#> [403] 0.52166931 0.61980309 0.64714336 0.74881096 0.59241177 0.85334619
#> [409] 0.08555970 0.66025436 0.31584485 0.89913719 0.60012276 0.51285681
#> [415] 0.54497719 0.37298837 0.77101359 0.50907600 0.63154747 0.66957774
#> [421] 0.19089558 0.90858198 0.71969388 0.83111022 0.20341560 0.50530687
#> [427] 0.22099096 0.84317090 0.62788977 0.49254746 0.54663895 0.83566363
#> [433] 0.58312658 0.64245717 0.86701495 0.08606837 0.55304387 0.47637028
#> [439] 0.30642376 0.77649413 0.24380597 0.63470371 0.56488311 0.18237467
#> [445] 0.32070193 0.64141563 0.90418390 0.66423866 0.43105908 0.59653266
#> [451] 0.62416583 0.60872593 0.36710610 0.58453011 0.45051941 0.72737963
#> [457] 0.30640042 0.32143009 0.49720056 0.80549411 0.48474239 0.60237841
#> [463] 0.19181671 0.67443561 0.51337010 0.28071554 0.27405884 0.81288504
#> [469] 0.92671462 0.87147419 0.89588382 0.87120086 0.88238202 0.88030072
#> [475] 0.81962405 0.83178488 0.91305674 0.77231617 0.88189740 0.33629192
#> [481] 0.61270148 0.67420464 0.48120673 0.59537346 0.90639347 0.79470993
#> [487] 0.53061932 0.79626649 0.84636294 0.76560014 0.21812773 0.85688837
#> [493] 0.83631537 0.70205075 0.56892174 0.80023955 0.63318115 0.84187505
#> [499] 0.12415413 0.88924238 0.39374736 0.76997436 0.49237370 0.71958259
#> [505] 0.40561818 0.70904981