Diversity index

Calculate the diversity (or entropy) index.

calcUncertaintyIndex(belongmatrix)

Arguments

belongmatrix

A membership matrix

Value

A vector with the values of the diversity (entropy) index

Details

The diversity (or entropy) index (Theil 1972) is calculated for each observation an varies between 0 and 1. When the value is close to 0, the observation belong to only one cluster (as in hard clustering). When the value is close to 1, the observation is undecided and tends to belong to each cluster. Values above 0.9 should be investigated. The formula is:

$$H2_{i} = \frac{-\sum[u_{ij}\ln(u_{ij})]}{\ln(k)}$$

with i and observation, j a cluster, k the number of clusters and u the membership matrix.

It is a simplified formula because the sum of each row of a membership matrix is 1.

References

Theil H (1972). Statistical decomposition analysis; with applications in the social and administrative sciences. North-Holland.

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- sf::st_drop_geometry(LyonIris[AnalysisFields])
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%[1] "criterion reached"
calcUncertaintyIndex(result$Belongings)
#>   [1] 0.66738812 0.22651401 0.30549241 0.67487450 0.22754735 0.20703113
#>   [7] 0.21742461 0.48979912 0.37386736 0.52331762 0.24319724 0.69497967
#>  [13] 0.43701101 0.69887398 0.68052206 0.49141748 0.54679161 0.68373693
#>  [19] 0.64487270 0.90830856 0.44697453 0.60276789 0.77498405 0.42201209
#>  [25] 0.29662341 0.43104662 0.83953974 0.65896159 0.29816204 0.50217330
#>  [31] 0.20445946 0.79274379 0.91153804 0.65958600 0.63903652 0.82824937
#>  [37] 0.24841469 0.39990422 0.53034034 0.91834765 0.63717141 0.28485685
#>  [43] 0.36929685 0.60898389 0.74620380 0.18875668 0.52214990 0.61693168
#>  [49] 0.69307469 0.59155142 0.36746594 0.44069041 0.87543253 0.28662707
#>  [55] 0.86302656 0.99490048 0.77641948 0.32843381 0.20525914 0.62391164
#>  [61] 0.41965728 0.52448290 0.17186990 0.94131550 0.48914042 0.21067573
#>  [67] 0.26049212 0.27147038 0.43786375 0.35158452 0.20866527 0.53877086
#>  [73] 0.62932032 0.13155916 0.45489883 0.65709387 0.34629975 0.46174144
#>  [79] 0.71977527 0.69993857 0.65853159 0.57245101 0.32058555 0.35354440
#>  [85] 0.42859037 0.44041540 0.55991128 0.20169419 0.65774387 0.84383866
#>  [91] 0.56624520 0.62012416 0.74761811 0.95732182 0.61916669 0.53185323
#>  [97] 0.61658965 0.89904086 0.71831902 0.90362768 0.29051906 0.46935157
#> [103] 0.57208709 0.40289033 0.96942200 0.69516800 0.68038804 0.25010089
#> [109] 0.35066342 0.82267084 0.63504998 0.80613000 0.13344362 0.66930495
#> [115] 0.16518756 0.58799383 0.66436229 0.61593597 0.34730975 0.13764364
#> [121] 0.80928104 0.51144500 0.38689637 0.14117057 0.71201365 0.77463531
#> [127] 0.93267569 0.60910067 0.59639181 0.22809385 0.50812691 0.57800721
#> [133] 0.20081988 0.73829728 0.31174653 0.39307834 0.50510867 0.68916440
#> [139] 0.17954135 0.36572552 0.18434128 0.61801544 0.74936629 0.29388672
#> [145] 0.59229500 0.70085672 0.71867316 0.39867966 0.35257957 0.41053599
#> [151] 0.59328737 0.18765995 0.86462852 0.50446794 0.44130726 0.84938458
#> [157] 0.38301734 0.71262020 0.68736982 0.49960011 0.42163224 0.47673152
#> [163] 0.25115151 0.55748430 0.19956053 0.47957312 0.87754720 0.59012958
#> [169] 0.60409697 0.24257355 0.32969931 0.70712580 0.67023254 0.48548305
#> [175] 0.80991476 0.68688639 0.63304946 0.79095018 0.79225127 0.26341048
#> [181] 0.21615613 0.67349603 0.67809888 0.77061384 0.70797267 0.48017687
#> [187] 0.61833964 0.73825338 0.68420233 0.95032858 0.39794254 0.88400708
#> [193] 0.41206420 0.74496335 0.53410160 0.34061578 0.14867726 0.81420664
#> [199] 0.23056853 0.43074847 0.36102694 0.82779378 0.89775305 0.26505163
#> [205] 0.64213467 0.62333521 0.83869384 0.69867991 0.78122220 0.18967179
#> [211] 0.77640006 0.41826258 0.75689212 0.50313814 0.86989199 0.58237046
#> [217] 0.66834521 0.49286421 0.26170470 0.14584218 0.12026115 0.64632839
#> [223] 0.14017711 0.35440997 0.63637400 0.23732590 0.20138183 0.78254175
#> [229] 0.63726326 0.21221584 0.50372652 0.39724545 0.74374844 0.63934179
#> [235] 0.58068469 0.35654443 0.45537773 0.40357930 0.63420302 0.68575072
#> [241] 0.45829804 0.29262242 0.15492390 0.20817910 0.93819194 0.59330362
#> [247] 0.57015543 0.60443288 0.11941877 0.87973734 0.58703032 0.60397930
#> [253] 0.66862323 0.69182926 0.81408315 0.54883966 0.83411034 0.82819773
#> [259] 0.46047676 0.60779529 0.64348357 0.05724412 0.41816360 0.53914121
#> [265] 0.54131644 0.69069013 0.85111664 0.19850386 0.24243284 0.30804540
#> [271] 0.51248653 0.36079308 0.25147898 0.83581805 0.34356663 0.31694753
#> [277] 0.53640824 0.07222104 0.23223844 0.73762716 0.69074741 0.48291462
#> [283] 0.78938428 0.48391362 0.31138422 0.17021293 0.27731269 0.36721286
#> [289] 0.78892391 0.86351982 0.73665450 0.75464733 0.53413934 0.67548236
#> [295] 0.75711170 0.20478939 0.88133919 0.52992156 0.32875557 0.32794680
#> [301] 0.78002333 0.60739032 0.91055600 0.74710688 0.57493366 0.26064821
#> [307] 0.55441082 0.35968286 0.24706360 0.12587904 0.53375855 0.74640254
#> [313] 0.51633989 0.57870002 0.23944531 0.21733262 0.17766865 0.63341778
#> [319] 0.85674530 0.35442633 0.34846663 0.04757293 0.65791758 0.57471272
#> [325] 0.56228820 0.65953045 0.66598183 0.17518003 0.44221605 0.74784437
#> [331] 0.52286042 0.34986091 0.39556454 0.02710102 0.55656337 0.84690792
#> [337] 0.80902560 0.74246342 0.54032503 0.85085408 0.26676980 0.68833940
#> [343] 0.27991446 0.24525232 0.48803174 0.22538274 0.94084397 0.55645181
#> [349] 0.41248412 0.88453621 0.44649461 0.09895459 0.69854541 0.36907814
#> [355] 0.18719451 0.60542985 0.56561481 0.62273885 0.76968572 0.72575710
#> [361] 0.47933846 0.20807797 0.42401235 0.56012173 0.86575595 0.74989391
#> [367] 0.48127525 0.45038773 0.64742089 0.07810620 0.76479799 0.87493950
#> [373] 0.38628716 0.21342286 0.56313381 0.84715732 0.93806792 0.61482471
#> [379] 0.37314267 0.58487289 0.64034938 0.55439246 0.25676066 0.22459928
#> [385] 0.60565538 0.43620963 0.72312028 0.38967585 0.71688788 0.69262345
#> [391] 0.64636726 0.62163625 0.33312773 0.59254915 0.52026276 0.80656505
#> [397] 0.86566784 0.78011872 0.49005911 0.63449602 0.64591728 0.54732063
#> [403] 0.52192371 0.61999091 0.64720250 0.74908110 0.59260896 0.85404415
#> [409] 0.08536702 0.66060433 0.31369611 0.89981599 0.60093377 0.51491092
#> [415] 0.54428405 0.37364541 0.77157821 0.50887248 0.63375569 0.66993639
#> [421] 0.19120120 0.90837721 0.71880663 0.83125227 0.20330274 0.50992446
#> [427] 0.22078480 0.84347538 0.62823863 0.49260482 0.54696598 0.83556084
#> [433] 0.58258601 0.64336008 0.86737177 0.08587078 0.55184598 0.47684228
#> [439] 0.30869617 0.77585282 0.24410615 0.63776643 0.56487916 0.18505299
#> [445] 0.32089317 0.63947129 0.90409776 0.66439763 0.42927376 0.59672877
#> [451] 0.62511325 0.60795997 0.36665391 0.58400113 0.45006870 0.72641939
#> [457] 0.30619155 0.32162742 0.49696040 0.80623718 0.48468099 0.60278838
#> [463] 0.19164339 0.67453988 0.51639758 0.28014125 0.27377804 0.81304082
#> [469] 0.92652384 0.87197057 0.89652147 0.87150307 0.88138642 0.88052543
#> [475] 0.81976643 0.83140717 0.91284965 0.77245451 0.88116393 0.33811242
#> [481] 0.61400262 0.67440035 0.48393653 0.59835461 0.90632921 0.79613457
#> [487] 0.53300317 0.79606394 0.84675343 0.76526968 0.22012548 0.85677585
#> [493] 0.83706549 0.70228122 0.56809104 0.79978755 0.63258027 0.84187496
#> [499] 0.12443602 0.89002039 0.39918306 0.77088611 0.49482377 0.71888588
#> [505] 0.41058812 0.70848440