Spatial diagnostic

Utility function to facilitate the spatial diagnostic of a classification

Calculate the following indicators: Moran I index (spdep::moranI) for each column of the membership matrix, Join count test (spdep::joincount.multi) for the most likely groups of each datapoint, Spatial consistency index (see function spConsistency) and the Elsa statistic (see function calcElsa). Note that if the FCMres object given was constructed with rasters, the joincount statistic is not calculated and no p-values are provided for the Moran I indices.

spatialDiag(
  object,
  nblistw = NULL,
  window = NULL,
  undecided = NULL,
  matdist = NULL,
  nrep = 50
)

Arguments

object

A FCMres object, typically obtained from functions CMeans, GCMeans, SFCMeans, SGFCMeans. Can also be a simple membership matrix.

nblistw

A list.w object describing the neighbours typically produced by the spdep package. Required if data is a dataframe, see the parameter window if you use a list of rasters as input. Can also be NULL if object is a FCMres object.

window

If rasters were used for the classification, the window must be specified instead of a list.w object. Can also be NULL if object is a FCMres object.

undecided

A float giving the threslhod to detect undecided observations. An observation is undecided if its maximum membership value is bellow this float. If null, no observations are undecided.

matdist

A matrix representing the dissimilarity between the clusters. The matrix must be squared and the diagonal must be filled with zeros.

nrep

An integer indicating the number of permutation to do to simulate the random distribution of the spatial inconsistency

Value

A named list with :

MoranValues : the moran I values for each column of the membership matrix (spdep::MoranI)

JoinCounts : the result of the join count test calculated with the most likely group for each datapoint (spdep::joincount.multi)

SpConsist : the mean value of the spatial consistency index (the lower, the better, see ?spConsistency for details)

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- sf::st_drop_geometry(LyonIris[AnalysisFields])
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%[1] "criterion reached"
spatialDiag(result, undecided=0.45, nrep=30)
#> $MoranValues
#>      MoranI pvalue   Cluster
#> 1 0.8834876  0.001 Cluster_1
#> 2 0.7702697  0.001 Cluster_2
#> 3 0.7873420  0.001 Cluster_3
#> 4 0.9261739  0.001 Cluster_4
#> 5 0.7750745  0.001 Cluster_5
#> 
#> $JoinCounts
#>                     Joincount  Expected  Variance  z-value
#> 1:1                  41.16488  10.20000   1.34948  26.6555
#> 2:2                  22.05238   4.24752   0.64251  22.2125
#> 3:3                  22.90952   4.37822   0.65991  22.8121
#> 4:4                  49.42143  12.98020   1.63215  28.5241
#> 5:5                  28.13130   7.23762   1.01751  20.7131
#> undecided:undecided  11.85408   4.78218   0.71304   8.3749
#> 2:1                   7.04167  13.33069   1.91931  -4.5395
#> 3:1                   0.25476  13.53267   1.94557  -9.5193
#> 3:2                   6.96845   8.75644   1.33472  -1.5476
#> 4:1                   0.24306  23.22772   3.10659 -13.0406
#> 4:2                   0.00000  15.02970   2.11676 -10.3303
#> 4:3                   0.00000  15.25743   2.14582 -10.4156
#> 5:1                   0.55536  17.37030   2.42838 -10.7904
#> 5:2                   2.39672  11.23960   1.66166  -6.8600
#> 5:3                   4.97694  11.40990   1.68429  -4.9568
#> 5:4                  10.39147  19.58416   2.68102  -5.6143
#> undecided:1          11.95346  14.13861   2.02383  -1.5360
#> undecided:2           7.52341   9.14851   1.38786  -1.3795
#> undecided:3           6.72569   9.28713   1.40668  -2.1597
#> undecided:4           4.81216  15.94059   2.23248  -7.4480
#> undecided:5          13.62327  11.92079   1.75172   1.2863
#> Jtot                 77.46641 209.17426   6.71734 -50.8175
#> 
#> $SpConsist
#> [1] 0.1618141
#> 
#> $SpConsistSamples
#>  [1] 0.1594721 0.1668309 0.1599693 0.1624430 0.1622276 0.1608443 0.1630786
#>  [8] 0.1659752 0.1607633 0.1619264 0.1648118 0.1612899 0.1607517 0.1650688
#> [15] 0.1605864 0.1595271 0.1582313 0.1567594 0.1606298 0.1605346 0.1631085
#> [22] 0.1627352 0.1673084 0.1623819 0.1595429 0.1599243 0.1640408 0.1630041
#> [29] 0.1589603 0.1616950
#> 
#> $Elsa
#>   [1] 0.05396049 0.00000000 0.08868233 0.00000000 0.00000000 0.00000000
#>   [7] 0.01265920 0.00000000 0.00000000 0.07902360 0.00000000 0.00000000
#>  [13] 0.04920087 0.13621637 0.04664519 0.02616155 0.12686155 0.14514735
#>  [19] 0.00000000 0.02579439 0.06235346 0.07972297 0.15161589 0.06576301
#>  [25] 0.00000000 0.01984429 0.33403480 0.09913157 0.00000000 0.00000000
#>  [31] 0.00000000 0.00000000 0.07972297 0.05396049 0.00000000 0.07409236
#>  [37] 0.00000000 0.00000000 0.00000000 0.28818314 0.11342038 0.00000000
#>  [43] 0.07391759 0.10202380 0.00000000 0.02616155 0.11608428 0.11342038
#>  [49] 0.13640135 0.05396049 0.00000000 0.00000000 0.31971156 0.07391759
#>  [55] 0.04920087 0.04664519 0.07391759 0.03347970 0.00000000 0.00000000
#>  [61] 0.00000000 0.09459448 0.00000000 0.05969321 0.08483035 0.03347970
#>  [67] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [73] 0.10533976 0.00000000 0.00000000 0.09881012 0.00000000 0.01048652
#>  [79] 0.07409236 0.04920087 0.00000000 0.07644196 0.05459018 0.00000000
#>  [85] 0.00000000 0.00000000 0.09913157 0.11999962 0.00000000 0.07739431
#>  [91] 0.00000000 0.05969321 0.00000000 0.00000000 0.05396049 0.00000000
#>  [97] 0.16058387 0.02579439 0.00000000 0.02088872 0.07739431 0.23216075
#> [103] 0.07391759 0.25151374 0.21453915 0.19653671 0.21772102 0.00000000
#> [109] 0.00000000 0.00000000 0.18523089 0.21584198 0.01131825 0.01341991
#> [115] 0.00000000 0.05969321 0.05828609 0.00000000 0.00000000 0.00000000
#> [121] 0.11342038 0.17294071 0.00000000 0.00000000 0.09459448 0.19260717
#> [127] 0.13594937 0.00000000 0.00000000 0.00000000 0.00000000 0.03497173
#> [133] 0.00000000 0.14514735 0.00000000 0.01401797 0.00000000 0.19653671
#> [139] 0.00000000 0.00000000 0.00000000 0.00000000 0.07391759 0.03347970
#> [145] 0.03972230 0.21772102 0.26897297 0.00000000 0.00000000 0.00000000
#> [151] 0.02616155 0.00000000 0.04216556 0.12684963 0.00000000 0.00000000
#> [157] 0.00000000 0.00000000 0.19498087 0.00000000 0.00000000 0.00000000
#> [163] 0.00000000 0.09913157 0.00000000 0.12074146 0.14514735 0.29273636
#> [169] 0.07972297 0.00000000 0.07972297 0.25151374 0.14514735 0.00000000
#> [175] 0.15451635 0.17013056 0.04157379 0.23446505 0.14715843 0.04664519
#> [181] 0.05789691 0.14514735 0.27455198 0.20868708 0.10202380 0.12686155
#> [187] 0.07972297 0.00000000 0.26119326 0.44673338 0.03837758 0.20498620
#> [193] 0.00000000 0.10125050 0.00000000 0.01692232 0.01285668 0.19900905
#> [199] 0.00000000 0.00000000 0.00000000 0.00000000 0.27438629 0.07144806
#> [205] 0.02579439 0.01285668 0.05636526 0.09881012 0.31971156 0.00000000
#> [211] 0.13251518 0.00000000 0.14672948 0.02152455 0.00000000 0.00000000
#> [217] 0.02152455 0.00000000 0.00000000 0.03844632 0.00000000 0.00000000
#> [223] 0.02652706 0.00000000 0.15172550 0.00000000 0.00000000 0.09881012
#> [229] 0.00000000 0.00000000 0.00000000 0.00000000 0.09913157 0.19232312
#> [235] 0.16140242 0.00000000 0.08156102 0.08868233 0.08156102 0.00000000
#> [241] 0.01692232 0.00000000 0.00000000 0.00000000 0.06949832 0.06081608
#> [247] 0.00000000 0.15391948 0.01632699 0.24253520 0.04863779 0.01285668
#> [253] 0.01285668 0.02088872 0.00000000 0.10874802 0.15916237 0.17296609
#> [259] 0.00000000 0.00000000 0.19368705 0.00000000 0.00000000 0.01285668
#> [265] 0.03837758 0.04763499 0.32934453 0.02088872 0.04216556 0.00000000
#> [271] 0.00000000 0.00000000 0.00000000 0.13342157 0.02088872 0.00000000
#> [277] 0.16170161 0.00000000 0.00000000 0.16140242 0.00000000 0.00000000
#> [283] 0.00000000 0.00000000 0.00000000 0.00000000 0.04216556 0.00000000
#> [289] 0.07610903 0.06949832 0.03469194 0.03469194 0.14944669 0.13305675
#> [295] 0.00000000 0.00000000 0.18091732 0.05139353 0.02652706 0.01632699
#> [301] 0.02652706 0.13791043 0.19609197 0.13997581 0.03837758 0.10878455
#> [307] 0.00000000 0.02088872 0.00000000 0.00000000 0.00000000 0.22382182
#> [313] 0.00000000 0.10670419 0.00000000 0.00000000 0.00000000 0.03837758
#> [319] 0.22546105 0.08079624 0.07391759 0.00000000 0.02086520 0.12163216
#> [325] 0.00000000 0.10874802 0.00000000 0.00000000 0.09331721 0.11003426
#> [331] 0.18281183 0.01285668 0.04763499 0.00000000 0.04763499 0.16590742
#> [337] 0.03581903 0.09331721 0.00000000 0.25151374 0.00000000 0.04863779
#> [343] 0.00000000 0.00862784 0.00000000 0.01041542 0.08982738 0.04763499
#> [349] 0.00000000 0.36276183 0.10202380 0.00000000 0.09331721 0.03179248
#> [355] 0.00000000 0.01560892 0.03837758 0.06262835 0.00000000 0.00000000
#> [361] 0.00000000 0.05636526 0.00000000 0.14601788 0.12175482 0.13652468
#> [367] 0.00000000 0.04863779 0.04763499 0.04131470 0.15161589 0.02575577
#> [373] 0.00000000 0.00000000 0.03497173 0.00000000 0.00000000 0.00000000
#> [379] 0.00000000 0.19529947 0.03469194 0.09971998 0.00000000 0.00000000
#> [385] 0.14492622 0.00000000 0.08115888 0.00000000 0.02652706 0.00000000
#> [391] 0.03497173 0.10255641 0.00000000 0.00000000 0.03469194 0.17668690
#> [397] 0.23565767 0.07972297 0.00000000 0.00000000 0.05701508 0.02088872
#> [403] 0.00000000 0.00000000 0.00000000 0.00000000 0.04763499 0.07688176
#> [409] 0.00000000 0.00000000 0.10004204 0.07610903 0.02152455 0.11342038
#> [415] 0.12353898 0.06969202 0.09331721 0.00000000 0.00000000 0.00000000
#> [421] 0.03347970 0.13140136 0.05139353 0.00000000 0.00000000 0.02086520
#> [427] 0.00000000 0.02820567 0.08156102 0.00000000 0.06081608 0.17854725
#> [433] 0.00000000 0.09934067 0.25922598 0.00000000 0.03497173 0.01632699
#> [439] 0.00000000 0.22252811 0.00000000 0.04312011 0.11287394 0.00000000
#> [445] 0.00000000 0.11342038 0.25149145 0.13251518 0.07982802 0.09461016
#> [451] 0.01632699 0.00000000 0.06772437 0.03497173 0.02088872 0.10657052
#> [457] 0.00000000 0.00000000 0.04763499 0.01984429 0.00000000 0.09331721
#> [463] 0.00000000 0.02652706 0.00000000 0.01041542 0.00000000 0.17651321
#> [469] 0.50995616 0.38773810 0.32393665 0.22839322 0.40287430 0.31716839
#> [475] 0.02953773 0.16140242 0.47884047 0.02616155 0.36482335 0.00000000
#> [481] 0.02616155 0.07391759 0.00000000 0.03270784 0.64182888 0.13340302
#> [487] 0.11236165 0.40392447 0.21453915 0.05636526 0.00000000 0.11866564
#> [493] 0.10203904 0.07860410 0.29075838 0.16866225 0.26554506 0.39447231
#> [499] 0.00000000 0.18694226 0.01265920 0.22345479 0.01984429 0.28950763
#> [505] 0.02616155 0.21152821
#>