Descriptive statistics by group

Calculate some descriptive statistics of each group

summarizeClusters(data, belongmatrix, weighted = TRUE, dec = 3, silent = TRUE)

Arguments

data
The original dataframe used for the classification

belongmatrix
A membership matrix

weighted
A boolean indicating if the summary statistics must use the membership matrix columns as weights (TRUE) or simply assign each observation to its most likely cluster and compute the statistics on each subset (FALSE)

dec
An integer indicating the number of digits to keep when rounding (default is 3)

silent
A boolean indicating if the results must be printed or silently returned

Value

A list of length k (the number of group). Each element of the list is a dataframe with summary statistics for the variables of data for each group

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- LyonIris@data[AnalysisFields]
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)

#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%[1] "criterion reached"

summarizeClusters(dataset, result$Belongings)

#> $Cluster_1
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   51.175 22.623 15.178 6.736      10.604  7.711  6.424   7.67      
#> Q10  52.44  25.817 15.781 8.483      12.423  9.132  7.737   8.879     
#> Q25  54.5   29.307 17.113 10.878     15.068  11.76  11.238  11.53     
#> Q50  57.366 32.27  18.396 15.129     17.729  15.678 15.03   14.43     
#> Q75  61.411 36.353 18.853 20.717     20.385  19.019 19.026  17.548    
#> Q90  63.988 39.372 19.306 30.184     23.289  22.155 24.094  22.202    
#> Q95  66.082 41.108 19.691 35.213     25.748  24.948 31.363  28.218    
#> Mean 57.9   32.354 17.862 16.893     17.715  15.523 15.828  15.376    
#> Std  4.825  6.163  1.488  8.583      5.119   5.616  7.809   7.087     
#>      Pct_brevet NivVieMed
#> Q5   9.917      15024.56 
#> Q10  12.007     15865.47 
#> Q25  16.738     18453.84 
#> Q50  22.078     20083    
#> Q75  29.667     22599.47 
#> Q90  37.11      25301.33 
#> Q95  42.413     27121.78 
#> Mean 23.659     20504.18 
#> Std  10.972     3923.102 
#> 
#> $Cluster_2
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   51.25  19.702 14.027 6.47       16.042  8.058  8.707   10.392    
#> Q10  52.176 21.062 14.511 7.746      18.236  9.056  12.395  12.723    
#> Q25  54.449 23.079 15.684 10.869     21.438  11.165 20.008  16.935    
#> Q50  56.802 25.796 16.381 14.128     24.354  13.899 26.582  22.775    
#> Q75  58.542 29.888 17.118 18.966     27.595  17.373 32.381  30.852    
#> Q90  62.568 34.26  18.541 25.498     30.753  20.822 38.113  34.208    
#> Q95  64.443 38.826 18.929 30.033     32.394  24.088 40.779  37.809    
#> Mean 56.838 26.764 16.405 15.657     24.231  14.43  25.816  23.423    
#> Std  4.158  6.059  1.451  7.57       6.243   5.197  9.809   9.678     
#>      Pct_brevet NivVieMed
#> Q5   19.702     12412.86 
#> Q10  23.999     12883.94 
#> Q25  32.06      14036.19 
#> Q50  38.266     15853.28 
#> Q75  45.084     18485.02 
#> Q90  49.391     21154.46 
#> Q95  53.211     23705.53 
#> Mean 37.943     16474.23 
#> Std  12.06      3567.886 
#> 
#> $Cluster_3
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   48.698 15.803 13.665 6.301      12.767  9.51   5.033   6.738     
#> Q10  50.43  18.648 13.863 7.608      14.85   11.34  6.021   7.719     
#> Q25  52.504 20.653 14.637 11.367     17.347  13.972 8.375   9.711     
#> Q50  54.72  23.887 15.415 15.533     19.662  17.651 11.581  12.444    
#> Q75  57.805 27.639 16.369 22.997     21.892  21.208 17.195  16.119    
#> Q90  61.104 32.66  18.095 31.708     24.884  24.374 23.326  21.602    
#> Q95  63.276 37.488 18.816 37.838     26.793  27.498 29.671  27.713    
#> Mean 55.108 24.737 15.632 17.998     19.506  17.699 13.653  13.759    
#> Std  4.541  6.195  1.559  9.568      4.746   5.786  7.999   7.443     
#>      Pct_brevet NivVieMed
#> Q5   12.313     15585.94 
#> Q10  15.087     16350    
#> Q25  19.917     19046.49 
#> Q50  24.958     21336.15 
#> Q75  31.915     23852.67 
#> Q90  38.876     26000.53 
#> Q95  45.055     28328.18 
#> Mean 26.582     21448.47 
#> Std  11.072     3975.251 
#> 
#> $Cluster_4
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   50.462 27.271 16.212 6.046      8.344   6.809  5.775   7.575     
#> Q10  51.346 28.733 17.07  7.212      10.185  8.428  6.442   8.411     
#> Q25  53.499 30.995 18.471 9.961      12.186  10.745 8.07    10.875    
#> Q50  55.429 35.438 19.231 13.444     14.613  13.443 11.289  12.828    
#> Q75  58.586 39.468 19.767 17.425     17.073  17.302 13.563  15.191    
#> Q90  60.943 43.053 20.108 25.433     19.504  20.717 17.646  18.318    
#> Q95  63.167 44.423 20.351 32.142     21.388  23.556 19.989  20.527    
#> Mean 56.017 35.278 18.82  14.972     14.548  14.04  11.76   13.345    
#> Std  4.145  6.265  1.41   7.915      4.589   5.32   5.898   5.667     
#>      Pct_brevet NivVieMed
#> Q5   7.23       18025.13 
#> Q10  8.324      19096.61 
#> Q25  10.104     21285.02 
#> Q50  13.248     23277.57 
#> Q75  16.574     25936.44 
#> Q90  23.317     28817.51 
#> Q95  30.38      31719.77 
#> Mean 14.852     23611.37 
#> Std  8.586      4071.16  
#> 
#> $Cluster_5
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   44.845 14.866 12.662 10.966     13.351  11.934 3.826   6.745     
#> Q10  46.018 15.818 12.923 14.075     14.56   13.29  4.344   7.22      
#> Q25  49.854 18.874 13.728 20.784     16.962  16.052 5.891   7.976     
#> Q50  52.309 22.031 14.552 28.546     18.828  19.023 7.551   9.723     
#> Q75  54.848 25.82  15.656 36.367     20.848  22.213 10.312  12.381    
#> Q90  58.191 29.372 16.645 41.942     22.443  26.809 14.426  15.438    
#> Q95  60.572 32.783 17.645 44.534     23.129  29.053 18.911  18.492    
#> Mean 52.281 22.539 14.761 27.924     18.604  19.353 8.875   10.912    
#> Std  4.8    5.74   1.567  10.409     3.805   5.702  5.533   6.035     
#>      Pct_brevet NivVieMed
#> Q5   10.135     19098.5  
#> Q10  11.673     20316.83 
#> Q25  14.716     22402    
#> Q50  18.392     24549.68 
#> Q75  23.314     28417.05 
#> Q90  29.812     30992.25 
#> Q95  33.413     34059.45 
#> Mean 19.877     25276.26 
#> Std  8.882      4514.339 
#>