Summary method for FCMres

Calculate some descriptive statistics of each group of a FCMres object

# S3 method for FCMres
summary(object, data = NULL, weighted = TRUE, dec = 3, silent = TRUE, ...)

Arguments

object

A FCMres object, typically obtained from functions CMeans, GCMeans, SFCMeans, SGFCMeans

data

A dataframe to use for the summary statistics instead of obj$data

weighted

A boolean indicating if the summary statistics must use the membership matrix columns as weights (TRUE) or simply assign each observation to its most likely cluster and compute the statistics on each subset (FALSE)

dec

An integer indicating the number of digits to keep when rounding (default is 3)

silent

A boolean indicating if the results must be printed or silently returned

...

Not used

Value

A list of length k (the number of group). Each element of the list is a dataframe with summary statistics for the variables of data for each group

Examples

data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- sf::st_drop_geometry(LyonIris[AnalysisFields])
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
#> [1] "Standardizing the data (set parameter to FALSE to avoid this step)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%[1] "criterion reached"
summary(result)
#> $Cluster_1
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   -0.98  -1.191 -1.351 -1.218     -0.612  -1.474 -0.679  -0.641    
#> Q10  -0.741 -1.008 -1.181 -1.1       -0.094  -1.257 -0.315  -0.329    
#> Q25  -0.276 -0.724 -0.569 -0.781     0.471   -0.865 0.58    0.243     
#> Q50  0.178  -0.419 -0.235 -0.465     1.033   -0.387 1.303   0.929     
#> Q75  0.589  0.095  0.08   0.017      1.595   0.192  1.965   1.961     
#> Q90  1.39   0.625  0.776  0.668      2.139   0.794  2.57    2.37      
#> Q95  1.697  1.21   0.969  1.078      2.421   1.279  2.894   2.858     
#> Mean 0.222  -0.279 -0.215 -0.303     1.007   -0.299 1.237   1.047     
#> Std  0.84   0.76   0.674  0.748      1.09    0.879  1.082   1.197     
#>      Pct_brevet NivVieMed
#> Q5   -0.397     -1.912   
#> Q10  -0.04      -1.84    
#> Q25  0.642      -1.609   
#> Q50  1.161      -1.208   
#> Q75  1.71       -0.683   
#> Q90  2.071      -0.135   
#> Q95  2.34       0.352    
#> Mean 1.149      -1.078   
#> Std  0.959      0.725    
#> 
#> $Cluster_2
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   -1.434 -1.648 -1.529 -1.236     -1.047  -1.213 -1.089  -1.008    
#> Q10  -1.073 -1.283 -1.371 -1.111     -0.727  -0.858 -0.965  -0.898    
#> Q25  -0.645 -1.019 -0.996 -0.719     -0.26   -0.381 -0.707  -0.684    
#> Q50  -0.18  -0.574 -0.644 -0.271     0.143   0.262  -0.35   -0.345    
#> Q75  0.434  -0.126 -0.183 0.558      0.565   0.858  0.245   0.104     
#> Q90  1.07   0.49   0.599  1.329      1.019   1.387  0.871   0.679     
#> Q95  1.557  1.104  0.946  1.939      1.402   1.915  1.617   1.46      
#> Mean -0.111 -0.486 -0.527 -0.03      0.145   0.278  -0.145  -0.172    
#> Std  0.919  0.788  0.73   0.959      0.818   0.989  0.857   0.897     
#>      Pct_brevet NivVieMed
#> Q5   -0.962     -1.262   
#> Q10  -0.734     -1.061   
#> Q25  -0.356     -0.518   
#> Q50  0.077      -0.02    
#> Q75  0.611      0.451    
#> Q90  1.161      0.869    
#> Q95  1.618      1.342    
#> Mean 0.187      -0.019   
#> Std  0.865      0.804    
#> 
#> $Cluster_3
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   -2.262 -1.772 -2.057 -0.822     -0.995  -0.808 -1.206  -1.003    
#> Q10  -1.998 -1.655 -1.819 -0.5       -0.706  -0.538 -1.13   -0.965    
#> Q25  -1.26  -1.275 -1.466 0.196      -0.269  -0.062 -0.954  -0.854    
#> Q50  -0.687 -0.86  -1.082 0.977      0.062   0.464  -0.772  -0.641    
#> Q75  -0.154 -0.41  -0.535 1.75       0.417   0.977  -0.464  -0.331    
#> Q90  0.525  0.075  -0.102 2.284      0.697   1.774  -0.006  0.081     
#> Q95  1.009  0.549  0.328  2.478      0.825   2.138  0.531   0.412     
#> Mean -0.688 -0.798 -0.964 0.916      0.027   0.52   -0.62   -0.487    
#> Std  0.982  0.733  0.735  1.035      0.67    0.965  0.614   0.752     
#>      Pct_brevet NivVieMed
#> Q5   -1.1       -0.586   
#> Q10  -0.969     -0.314   
#> Q25  -0.7       0.121    
#> Q50  -0.408     0.556    
#> Q75  -0.012     1.347    
#> Q90  0.501      1.872    
#> Q95  0.785      2.411    
#> Mean -0.289     0.715    
#> Std  0.713      0.924    
#> 
#> $Cluster_4
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   -1.176 -0.301 -0.364 -1.322     -1.809  -1.652 -0.966  -0.914    
#> Q10  -0.883 -0.021 0.119  -1.17      -1.471  -1.334 -0.892  -0.802    
#> Q25  -0.439 0.275  0.768  -0.877     -1.107  -0.965 -0.708  -0.49     
#> Q50  -0.058 0.845  1.084  -0.528     -0.671  -0.484 -0.352  -0.249    
#> Q75  0.6    1.358  1.385  -0.122     -0.237  0.15   -0.098  0.037     
#> Q90  1.068  1.626  1.545  0.645      0.185   0.747  0.345   0.428     
#> Q95  1.506  1.993  1.613  1.317      0.502   1.186  0.608   0.709     
#> Mean 0.075  0.825  0.942  -0.373     -0.681  -0.377 -0.295  -0.184    
#> Std  0.836  0.794  0.654  0.779      0.801   0.896  0.648   0.695     
#>      Pct_brevet NivVieMed
#> Q5   -1.3       -0.831   
#> Q10  -1.211     -0.558   
#> Q25  -1.079     -0.141   
#> Q50  -0.815     0.276    
#> Q75  -0.551     0.832    
#> Q90  -0.017     1.422    
#> Q95  0.537      1.96     
#> Mean -0.685     0.357    
#> Std  0.679      0.822    
#> 
#> $Cluster_5
#>      Lden   NO2    PM25   VegHautPrt Pct0_14 Pct_65 Pct_Img TxChom1564
#> Q5   -0.907 -0.846 -0.823 -1.202     -1.394  -1.497 -0.901  -0.902    
#> Q10  -0.642 -0.44  -0.532 -1.046     -1.06   -1.208 -0.741  -0.768    
#> Q25  -0.211 0.077  0.134  -0.801     -0.58   -0.758 -0.348  -0.409    
#> Q50  0.451  0.458  0.729  -0.362     -0.085  -0.088 0.085   -0.046    
#> Q75  1.229  0.98   0.945  0.182      0.352   0.469  0.561   0.376     
#> Q90  1.731  1.361  1.168  1.107      0.88    0.989  1.167   0.965     
#> Q95  2.156  1.569  1.354  1.583      1.386   1.493  1.965   1.696     
#> Mean 0.512  0.471  0.489  -0.194     -0.095  -0.116 0.19    0.097     
#> Std  0.981  0.793  0.696  0.841      0.911   0.946  0.882   0.895     
#>      Pct_brevet NivVieMed
#> Q5   -1.103     -1.479   
#> Q10  -0.912     -1.254   
#> Q25  -0.483     -0.702   
#> Q50  -0.074     -0.412   
#> Q75  0.547      0.117    
#> Q90  1.14       0.607    
#> Q95  1.602      1.001    
#> Mean 0.067      -0.304   
#> Std  0.887      0.797    
#>